Function to extract a summary of the specified serological model object

This function extracts a summary corresponding to a serological model object containing information about the original serological survey data used to fit the model, such as the year when the survey took place, the type of test taken and the corresponding antibody, as well as information about the convergence of the model, like the expected log pointwise predictive density elpd and its corresponding standard deviation.

Usage

extract_seromodel_summary(seromodel_object, serodata)

Arguments

seromodel_object

Stanfit object containing the results of fitting a model by means of run_seromodel.

serodata

A data frame containing the data from a seroprevalence survey. This data frame must contain at least the following columns:

total

Number of samples for each age group

counts

Number of positive samples for each age group

tsur

Year in which the survey took place

age_mean_f

Floor value of the average between age_min and age_max

sample_size

The size of the sample

birth_year

The year in which the individuals of each age group were born

The last six columns can be added to serodata by means of the function prepare_serodata().

Value

model_summary. Object with a summary of seromodel_object containing the following:

foi_model

Name of the selected model.

data_set

Seroprevalence survey label

country

Name of the country were the survey was conducted in.

year

Year in which the survey was conducted.

test

Type of test of the survey.

antibody

Antibody

n_sample

Total number of samples in the survey.

n_agec

Number of age groups considered.

n_iter

Number of iterations by chain including warmup.

elpd

elpd

se

se

converged

convergence

Examples

data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- fit_seromodel(
  serodata = serodata,
  foi_model = "constant"
)
extract_seromodel_summary(seromodel_object,
  serodata = serodata
)
#>   foi_model    dataset country year  test         antibody n_sample n_agec
#> 1  constant COL-035-93     COL 2012 ELISA IgG anti-T.cruzi      747     72
#>   n_iter   elpd   se converged
#> 1   1000 -92.73 6.42       Yes