Fit Model

Function that fits the selected model to the data

Usage

fit_model(
  model_data,
  model_name,
  n_iters = 1000,
  n_thin = 2,
  delta = 0.9,
  m_treed = 10,
  decades = 0
)

Arguments

model_data

A data frame containing the data from a seroprevalence survey. For further details refer to run_model.

model_name

Name of the selected model. Current version provides three options:

constant_foi_bi

Runs a constant model

continuous_foi_normal_bi

Runs a normal model

continuous_foi_normal_log

Runs a normal logarithmic model

n_iters

Number of interations for eah chain including the warmup. iter in sampling.

n_thin

Positive integer specifying the period for saving samples. thin in sampling.

delta

Real number between 0 and 1 that represents the target average acceptance probability. Increasing the value of delta will result in a smaller step size and fewer divergences. For further details refer to the control parameter in sampling or here.

m_treed

Maximum tree depth for the binary tree used in the NUTS stan sampler. For further details refer to the control parameter in sampling.

decades

Number of decades covered by the survey data.

Value

model_object. An object containing relevant information about the implementation of the model. It contains the following:

fit	stanfit object returned by the function sampling

model_data	A data frame containing the data from a seroprevalence survey. For further details refer to run_model.

stan_data	List containing Nobs, Npos, Ntotal, Age, Ymax, AgeExpoMatrix and NDecades. This object is used as an input for the sampling function

real_exposure_years	Integer atomic vector containing the actual exposure years (1946, ..., 2007 e.g.)

exposure_years	Integer atomic vector containing the numeration of the exposure years.

n_iters	Number of interations for eah chain including the warmup.

n_thin	Positive integer specifying the period for saving samples.

n_warmup	Number of warm up iterations. Set by default as n_iters/2.

model_name	The name of the model

delta	Real number between 0 and 1 that represents the target average acceptance probability.

m_treed	Maximum tree depth for the binary tree used in the NUTS stan sampler.

loo_fit	Efficient approximate leave-one-out cross-validation. Refer to loo for further details.

foi_cent_est	A data fram e containing year (corresponding to real_exposure_years), lower, upper, and medianv

foi_post_s	Sample n rows from a table. Refer to sample_n for further details.

model_summary	A data fram containing the summary of the model. Refer to extract_model_summary for further details.

Examples

model_data <- prepare_data(mydata)
fit_model (model_data,
           model_name = "constant_foi_bi")
#> Warning: Model constant_foi_bi is being compiled for the first time. This might take some minutes
#> Warning: path[1]="inst/extdata/stanmodels/constant_foi_bi.stan": No such file or directory
#> Warning: cannot open file 'inst/extdata/stanmodels/constant_foi_bi.stan': No such file or directory
#> Error in file(fname, "rt") : cannot open the connection
#> Error in get_model_strcode(file, model_code): cannot open model file "inst/extdata/stanmodels/constant_foi_bi.stan"