Generate data frame containing the age of each cohort corresponding to each birth year excluding the year of the survey.

This function generates a data frame containing the age of each cohort corresponding to each birth_year excluding the year of the survey, for which the cohort age is still 0. specified serological survey data serodata excluding the year of the survey.

Usage

get_cohort_ages(serodata)

Arguments

serodata

A data frame containing the data from a seroprevalence survey. This data frame must contain at least the following columns:

total

Number of samples for each age group

counts

Number of positive samples for each age group

tsur

Year in which the survey took place

age_mean_f

Floor value of the average between age_min and age_max

sample_size

The size of the sample

birth_year

The year in which the individuals of each age group were born

The last six columns can be added to serodata by means of the function prepare_serodata().

Value

cohort_ages. A data frame containing the age of each cohort corresponding to each birth year

Examples

data(chagas2012)
serodata <- prepare_serodata(serodata = chagas2012, alpha = 0.05)
cohort_ages <- get_cohort_ages(serodata = serodata)