Generate data frame containing the confidence interval based on a force-of-infection fitting

This function computes the corresponding binomial confidence intervals for the obtained prevalence based on a fitting of the force-of-infection foi for plotting an analysis purposes.

Usage

get_prev_expanded(foi, serodata, alpha = 0.05, bin_data = FALSE, bin_step = 5)

Arguments

foi

Object containing the information of the force-of-infection. It is obtained from rstan::extract(seromodel_object$seromodel, "foi", inc_warmup = FALSE)[[1]].

serodata

A data frame containing the data from a seroprevalence survey. This data frame must contain at least the following columns:

total

Number of samples for each age group

counts

Number of positive samples for each age group

tsur

Year in which the survey took place

age_mean_f

Floor value of the average between age_min and age_max

sample_size

The size of the sample

birth_year

The year in which the individuals of each age group were born

The last six columns can be added to serodata by means of the function prepare_serodata().

alpha

Probability threshold for statistical significance used for both the binomial confidence interval, and the lower and upper quantiles of the estimated prevalence.

bin_data

If TRUE, serodata is binned by means of prepare_bin_data. Otherwise, age groups are kept as originally input.

bin_step

Integer specifying the age groups bin size to be used when bin_data is set to TRUE.

Value

prev_final. The expanded prevalence data. This is used for plotting purposes in the visualization module.

Examples

data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- run_seromodel(
  serodata = serodata,
  foi_model = "constant"
)
#> serofoi model constant finished running ------
#>           [,1]              
#> foi_model "constant"        
#> dataset   "COL-035-93"      
#> country   "COL"             
#> year      "2012"            
#> test      "ELISA"           
#> antibody  "IgG anti-T.cruzi"
#> n_sample  "747"             
#> n_agec    "72"              
#> n_iter    "1000"            
#> elpd      "-92.75"          
#> se        "6.4"             
#> converged "Yes"             
foi <- rstan::extract(seromodel_object, "foi")[[1]]
get_prev_expanded(foi, serodata)
#>    age predicted_prev predicted_prev_lower predicted_prev_upper   prev_obs
#> 1    1    0.003893214          0.003077191          0.004890482 0.00000000
#> 2    2    0.007771271          0.006144913          0.009757047 0.00000000
#> 3    3    0.011634230          0.009203195          0.014599812 0.02857143
#> 4    4    0.015482149          0.012252066          0.019418894 0.00000000
#> 5    5    0.019315088          0.015291555          0.024214408 0.00000000
#> 6    6    0.023133104          0.018321690          0.028986470 0.00000000
#> 7    7    0.026936256          0.021342502          0.033735194 0.00000000
#> 8    8    0.030724601          0.024354018          0.038460694 0.00000000
#> 9    9    0.034498198          0.027356267          0.043163085 0.00000000
#> 10  10    0.038257103          0.030349277          0.047842478 0.00000000
#> 11  11    0.042001374          0.033333078          0.052498988 0.07692308
#> 12  12    0.045731067          0.036307697          0.057132724 0.00000000
#> 13  13    0.049446241          0.039273162          0.061743799 0.00000000
#> 14  14    0.053146950          0.042229502          0.066332324 0.00000000
#> 15  15    0.056833251          0.045176744          0.070898409 0.00000000
#> 16  16    0.060505201          0.048114918          0.075442164 0.00000000
#> 17  17    0.064162855          0.051044050          0.079963697 0.00000000
#> 18  18    0.067806270          0.053964169          0.084463118 0.00000000
#> 19  19    0.071435499          0.056875301          0.088940534 0.00000000
#> 20  20    0.075050600          0.059777476          0.093396054 0.00000000
#> 21  21    0.078651626          0.062670720          0.097829784 0.00000000
#> 22  22    0.082238632          0.065555061          0.102241831 0.00000000
#> 23  23    0.085811673          0.068430527          0.106632301 0.00000000
#> 24  24    0.089370804          0.071297144          0.111001300 0.00000000
#> 25  25    0.092916078          0.074154940          0.115348932 0.00000000
#> 26  26    0.096447550          0.077003942          0.119675302 0.09090909
#> 27  27    0.099965273          0.079844177          0.123980514 0.12500000
#> 28  28    0.103469301          0.082675672          0.128264671 0.00000000
#> 29  29    0.106959687          0.085498454          0.132527877 0.00000000
#> 30  30    0.110436484          0.088312550          0.136770234 0.10000000
#> 31  31    0.113899745          0.091117986          0.140991843 0.10000000
#> 32  32    0.117349523          0.093914790          0.145192807 0.00000000
#> 33  33    0.120785870          0.096702987          0.149373226 0.09090909
#> 34  34    0.124208839          0.099482604          0.153533201 0.00000000
#> 35  35    0.127618481          0.102253668          0.157672832 0.10000000
#> 36  36    0.131014849          0.105016205          0.161792217 0.30000000
#> 37  37    0.134397994          0.107770241          0.165891457 0.11111111
#> 38  38    0.137767968          0.110515802          0.169970650 0.25000000
#> 39  39    0.141124822          0.113252915          0.174029893 0.00000000
#> 40  40    0.144468607          0.115981605          0.178069285 0.22222222
#> 41  41    0.147799374          0.118701898          0.182088922 0.50000000
#> 42  42    0.151117173          0.121413821          0.186088902 0.00000000
#> 43  43    0.154422055          0.124117398          0.190069319 0.33333333
#> 44  44    0.157714071          0.126812656          0.194030271 1.00000000
#> 45  45    0.160993271          0.129499620          0.197971851 0.00000000
#> 46  46    0.164259703          0.132178316          0.201894155 0.66666667
#> 47  47    0.167513419          0.134848769          0.205797277         NA
#> 48  48    0.170754468          0.137511004          0.209681311 1.00000000
#> 49  49    0.173982898          0.140165048          0.213546350 0.66666667
#> 50  50    0.177198759          0.142810924          0.217392488 0.00000000
#> 51  51    0.180402101          0.145448658          0.221219815 1.00000000
#> 52  52    0.183592971          0.148078276          0.225028426 0.50000000
#> 53  53    0.186771418          0.150699802          0.228818410 0.00000000
#> 54  54    0.189937491          0.153313260          0.232589860 0.33333333
#> 55  55    0.193091237          0.155918677          0.236342865 0.50000000
#> 56  56    0.196232706          0.158516076          0.240077517 0.00000000
#> 57  57    0.199361944          0.161105483          0.243793904 0.75000000
#> 58  58    0.202478999          0.163686922          0.247492116 0.66666667
#> 59  59    0.205583919          0.166260416          0.251172242 0.00000000
#> 60  60    0.208676751          0.168825992          0.254834371 0.09090909
#> 61  61    0.211757542          0.171383673          0.258478590 0.00000000
#> 62  62    0.214826338          0.173933484          0.262104987 1.00000000
#> 63  63    0.217883187          0.176475448          0.265713649 0.50000000
#> 64  64    0.220928135          0.179009590          0.269304663 0.60000000
#> 65  65    0.223961229          0.181535935          0.272878115 0.25000000
#> 66  66    0.226982514          0.184054505          0.276434092 1.00000000
#> 67  67    0.229992036          0.186565325          0.279972678 1.00000000
#> 68  68    0.232989842          0.189068418          0.283493958 0.00000000
#> 69  69    0.235975977          0.191563810          0.286998018 0.40000000
#> 70  70    0.238950486          0.194051522          0.290484941 0.66666667
#> 71  71    0.241913414          0.196531579          0.293954812 1.00000000
#> 72  72    0.244864808          0.199004005          0.297407713 1.00000000
#> 73  73    0.247804710          0.201468822          0.300843728         NA
#> 74  74    0.250733168          0.203926055          0.304262939         NA
#> 75  75    0.253650224          0.206375727          0.307665428         NA
#> 76  76    0.256555923          0.208817860          0.311051278         NA
#> 77  77    0.259450310          0.211252478          0.314420569 1.00000000
#>    prev_obs_lower prev_obs_upper total counts     survey
#> 1    0.0000000000     0.10281792    34      0 COL-035-93
#> 2    0.0000000000     0.13718517    25      0 COL-035-93
#> 3    0.0007231044     0.14917208    35      1 COL-035-93
#> 4    0.0000000000     0.11944487    29      0 COL-035-93
#> 5    0.0000000000     0.09739376    36      0 COL-035-93
#> 6    0.0000000000     0.14818513    23      0 COL-035-93
#> 7    0.0000000000     0.10281792    34      0 COL-035-93
#> 8    0.0000000000     0.18530197    18      0 COL-035-93
#> 9    0.0000000000     0.16109762    21      0 COL-035-93
#> 10   0.0000000000     0.14247360    24      0 COL-035-93
#> 11   0.0019456285     0.36029744    13      1 COL-035-93
#> 12   0.0000000000     0.12343612    28      0 COL-035-93
#> 13   0.0000000000     0.21801936    15      0 COL-035-93
#> 14   0.0000000000     0.13718517    25      0 COL-035-93
#> 15   0.0000000000     0.24705264    13      0 COL-035-93
#> 16   0.0000000000     0.20590721    16      0 COL-035-93
#> 17   0.0000000000     0.26464847    12      0 COL-035-93
#> 18   0.0000000000     0.18530197    18      0 COL-035-93
#> 19   0.0000000000     0.33626712     9      0 COL-035-93
#> 20   0.0000000000     0.36941665     8      0 COL-035-93
#> 21   0.0000000000     0.60236464     4      0 COL-035-93
#> 22   0.0000000000     0.17646691    19      0 COL-035-93
#> 23   0.0000000000     0.21801936    15      0 COL-035-93
#> 24   0.0000000000     0.23163576    14      0 COL-035-93
#> 25   0.0000000000     0.16843347    20      0 COL-035-93
#> 26   0.0022989722     0.41277992    11      1 COL-035-93
#> 27   0.0031597235     0.52650967     8      1 COL-035-93
#> 28   0.0000000000     0.33626712     9      0 COL-035-93
#> 29   0.0000000000     0.33626712     9      0 COL-035-93
#> 30   0.0025285785     0.44501612    10      1 COL-035-93
#> 31   0.0025285785     0.44501612    10      1 COL-035-93
#> 32   0.0000000000     0.28491415    11      0 COL-035-93
#> 33   0.0022989722     0.41277992    11      1 COL-035-93
#> 34   0.0000000000     0.36941665     8      0 COL-035-93
#> 35   0.0025285785     0.44501612    10      1 COL-035-93
#> 36   0.0667395112     0.65245285    10      3 COL-035-93
#> 37   0.0028091367     0.48249651     9      1 COL-035-93
#> 38   0.0063094632     0.80587955     4      1 COL-035-93
#> 39   0.0000000000     0.60236464     4      0 COL-035-93
#> 40   0.0281449735     0.60009357     9      2 COL-035-93
#> 41   0.0675859865     0.93241401     4      2 COL-035-93
#> 42   0.0000000000     0.36941665     8      0 COL-035-93
#> 43   0.0084037587     0.90570068     3      1 COL-035-93
#> 44   0.0250000000     1.00000000     1      1 COL-035-93
#> 45   0.0000000000     0.70759823     3      0 COL-035-93
#> 46   0.0942993241     0.99159624     3      2 COL-035-93
#> 47             NA             NA    NA     NA COL-035-93
#> 48   0.1581138830     1.00000000     2      2 COL-035-93
#> 49   0.0942993241     0.99159624     3      2 COL-035-93
#> 50   0.0000000000     0.70759823     3      0 COL-035-93
#> 51   0.0250000000     1.00000000     1      1 COL-035-93
#> 52   0.1181172488     0.88188275     6      3 COL-035-93
#> 53   0.0000000000     0.60236464     4      0 COL-035-93
#> 54   0.0084037587     0.90570068     3      1 COL-035-93
#> 55   0.1181172488     0.88188275     6      3 COL-035-93
#> 56   0.0000000000     0.97500000     1      0 COL-035-93
#> 57   0.1941204497     0.99369054     4      3 COL-035-93
#> 58   0.0942993241     0.99159624     3      2 COL-035-93
#> 59   0.0000000000     0.97500000     1      0 COL-035-93
#> 60   0.0022989722     0.41277992    11      1 COL-035-93
#> 61   0.0000000000     0.52182375     5      0 COL-035-93
#> 62   0.0250000000     1.00000000     1      1 COL-035-93
#> 63   0.0125791171     0.98742088     2      1 COL-035-93
#> 64   0.1466327996     0.94725505     5      3 COL-035-93
#> 65   0.0318540262     0.65085579     8      2 COL-035-93
#> 66   0.0250000000     1.00000000     1      1 COL-035-93
#> 67   0.0250000000     1.00000000     1      1 COL-035-93
#> 68   0.0000000000     0.84188612     2      0 COL-035-93
#> 69   0.0527449505     0.85336720     5      2 COL-035-93
#> 70   0.0942993241     0.99159624     3      2 COL-035-93
#> 71   0.0250000000     1.00000000     1      1 COL-035-93
#> 72   0.0250000000     1.00000000     1      1 COL-035-93
#> 73             NA             NA    NA     NA COL-035-93
#> 74             NA             NA    NA     NA COL-035-93
#> 75             NA             NA    NA     NA COL-035-93
#> 76             NA             NA    NA     NA COL-035-93
#> 77   0.0250000000     1.00000000     1      1 COL-035-93