Generate probabilities of being previously exposed to a pathogen given a historical force-of-infection.

Generate probabilities of being previously exposed to a pathogen given a historical force-of-infection.

Usage

get_sim_probability(sim_data, foi, mu = 0, model_type = "time-varying")

Arguments

sim_data

A dataframe object containing the following columns:

age

Age group markers

tsur

Year of the survey

foi

Numeric atomic vector corresponding to the desired time-varying or age-varying Force-of-Infection to simulate from

mu

Seroreversion rate

model_type

String specifying the type of model to be used. Current valid options are 'time-varying' and 'age-varying'

Value

A dataframe containing the following columns:

age

Exposure ages

probability

Probability to obtain a seropositive case for each age according to the provided FoI

Examples

n_years <- 50
sim_data <- data.frame(
  age = seq(1,n_years),
  tsur = 2050
)
foi <- rep(0.02, n_years)
sim_probability <- get_sim_probability(sim_data = sim_data, foi=foi)