Plots model summary

Plots model summary

Usage

plot_summary(
  seromodel,
  serosurvey,
  loo_estimate_digits = 1,
  central_estimate_digits = 2,
  rhat_digits = 2,
  size_text = 11
)

Arguments

seromodel

stan_fit object obtained from sampling a model with fit_seromode

serosurvey

survey_year

Year in which the survey took place

age_min

Floor value of the average between age_min and age_max

age_max

The size of the sample

sample_size

Number of samples for each age group

n_seropositive

Number of positive samples for each age group

loo_estimate_digits

Number of loo estimate digits

central_estimate_digits

Number of central estimate digits

size_text

Size of text for plotting (base_size in ggplot2)

Value

ggplot object with a summary of the specified model