Function that prepares a pre-processed serological survey dataset to plot the binomial confidence intervals of the seroprevalence grouped by age group

This function prepapares a given pre-processed serological dataset (see prepare_serodata) to plot the binomial confidence intervals of its corresponding seroprevalence grouped by age group.

Usage

prepare_bin_data(serodata)

Arguments

serodata

A data frame containing the data from a seroprevalence survey. This data frame must contain the following columns:

survey	survey Label of the current survey

total	Number of samples for each age group

counts	Number of positive samples for each age group

age_min	age_min

age_max	age_max

tsur	Year in which the survey took place

country	The country where the survey took place

test	The type of test taken

antibody	antibody

age_mean_f	Floor value of the average between age_min and age_max

sample_size	The size of the sample

birth_year	The year in which the individuals of each age group were bornt

prev_obs	Observed prevalence

prev_obs_lower	Lower limit of the confidence interval for the observed prevalence

prev_obs_upper	Upper limit of the confidence interval for the observed prevalence

The last six colums can be added to serodata by means of the function prepare_serodata.

Value

data set with the binomial confidence intervals

Examples

data(chagas2012)
serodata <- prepare_serodata(chagas2012)
prepare_bin_data(serodata)
#>      cut_ages bin_size bin_pos age   p_obs_bin  p_obs_bin_l p_obs_bin_u
#> X       [1,6]      182       1   4 0.005494505 0.0001390992  0.03023169
#> X.1    (6,11]      110       1   8 0.009090909 0.0002301354  0.04961119
#> X.2   (11,16]       97       0  14 0.000000000 0.0000000000  0.03731564
#> X.3   (16,21]       51       0  18 0.000000000 0.0000000000  0.06977703
#> X.4   (21,26]       79       1  24 0.012658228 0.0003204272  0.06851991
#> X.5   (26,31]       46       3  28 0.065217391 0.0136567719  0.17896439
#> X.6   (31,36]       50       5  34 0.100000000 0.0332750936  0.21813537
#> X.7   (36,41]       30       6  38 0.200000000 0.0771355120  0.38566651
#> X.8   (41,46]       18       4  44 0.222222222 0.0640920477  0.47637277
#> X.9   (46,51]        9       5  48 0.555555556 0.2120085068  0.86300434
#> X.10  (51,56]       20       7  54 0.350000000 0.1539092048  0.59218853
#> X.11  (56,61]       24       6  58 0.250000000 0.0977304095  0.46711280
#> X.12  (61,66]       17       8  64 0.470588235 0.2298326873  0.72188170
#> X.13  (66,71]       12       6  68 0.500000000 0.2109446382  0.78905536
#> X.14  (71,76]        1       1  74 1.000000000 0.0250000000  1.00000000
#> X.15  (76,81]        1       1  78 1.000000000 0.0250000000  1.00000000