Run specified stan model for the force-of-infection and estimate the seroprevalence based on the result of the fit

This function runs the specified model for the force-of-infection foi_model using the data from a seroprevalence survey serodata as the input data. See fit_seromodel for further details.

Usage

run_seromodel(
  serodata,
  foi_model = c("constant", "tv_normal_log", "tv_normal"),
  foi_location = 0,
  foi_scale = 1,
  chunk_size = 1,
  chunks = NULL,
  iter = 1000,
  adapt_delta = 0.9,
  max_treedepth = 10,
  seed = 12345,
  print_summary = TRUE,
  ...
)

Arguments

serodata

A data frame containing the data from a seroprevalence survey. This data frame must contain at least the following columns:

total

Number of samples for each age group

counts

Number of positive samples for each age group

tsur

Year in which the survey took place

age_mean_f

Floor value of the average between age_min and age_max

sample_size

The size of the sample

birth_year

The year in which the individuals of each age group were born

The last six columns can be added to serodata by means of the function prepare_serodata().

foi_model

Name of the selected model. Current version provides three options:

"constant"

Runs a constant model

"tv_normal"

Runs a normal model

"tv_normal_log"

Runs a normal logarithmic model

foi_location

Location parameter of the force-of-infection distribution of the selected model. Depending on foi_model, the meaning may vary.

foi_scale

Scale parameter of the force-of-infection distribution of the selected model. Depending on foi_model, the meaning may vary.

chunk_size

Size of the chunks to be used in case that the chunk structure chunks is not specified in fit_seromodel. Default is 1, meaning that one force of infection value is to be estimated for every year in the time interval spanned by the serosurvey. If the length of the time interval is not exactly divisible by chunk_size, the remainder years are included in the last chunk.

chunks

Numeric list specifying the chunk structure of the time interval from the birth year of the oldest age cohort min(serodata$age_mean_f) to the time when the serosurvey was conducted t_sur. If NULL, the time interval is divided in chunks of size chunk_size.

iter

Number of interactions for each chain including the warmup. iter in sampling.

adapt_delta

Real number between 0 and 1 that represents the target average acceptance probability. Increasing the value of adapt_delta will result in a smaller step size and fewer divergences. For further details refer to the control parameter in sampling or here.

max_treedepth

Maximum tree depth for the binary tree used in the NUTS stan sampler. For further details refer to the control parameter in sampling.

seed

For further details refer to the seed parameter in sampling.

print_summary

Boolean. If TRUE, a table summarizing modelling results is printed.

...

Additional parameters for sampling.

Value

seromodel_object. An object containing relevant information about the implementation of the model. For further details refer to fit_seromodel.

Examples

data(chagas2012)
serodata <- prepare_serodata(chagas2012)
run_seromodel(
  serodata,
  foi_model = "constant"
)
#> serofoi model constant finished running ------
#>           [,1]              
#> foi_model "constant"        
#> dataset   "COL-035-93"      
#> country   "COL"             
#> year      "2012"            
#> test      "ELISA"           
#> antibody  "IgG anti-T.cruzi"
#> n_sample  "747"             
#> n_agec    "72"              
#> n_iter    "1000"            
#> elpd      "-92.73"          
#> se        "6.42"            
#> converged "Yes"             
#> Inference for Stan model: constant.
#> 4 chains, each with iter=1000; warmup=500; thin=1; 
#> post-warmup draws per chain=500, total post-warmup draws=2000.
#> 
#>                           mean se_mean   sd    2.5%     25%     50%     75%
#> lambda0                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> prob_infected[1]          0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> prob_infected[2]          0.01    0.00 0.00    0.01    0.01    0.01    0.01
#> prob_infected[3]          0.01    0.00 0.00    0.01    0.01    0.01    0.01
#> prob_infected[4]          0.02    0.00 0.00    0.01    0.01    0.02    0.02
#> prob_infected[5]          0.02    0.00 0.00    0.01    0.02    0.02    0.02
#> prob_infected[6]          0.02    0.00 0.00    0.02    0.02    0.02    0.03
#> prob_infected[7]          0.03    0.00 0.00    0.02    0.02    0.03    0.03
#> prob_infected[8]          0.03    0.00 0.00    0.02    0.03    0.03    0.03
#> prob_infected[9]          0.03    0.00 0.00    0.03    0.03    0.03    0.04
#> prob_infected[10]         0.04    0.00 0.01    0.03    0.03    0.04    0.04
#> prob_infected[11]         0.04    0.00 0.01    0.03    0.04    0.04    0.05
#> prob_infected[12]         0.05    0.00 0.01    0.03    0.04    0.05    0.05
#> prob_infected[13]         0.05    0.00 0.01    0.04    0.04    0.05    0.05
#> prob_infected[14]         0.05    0.00 0.01    0.04    0.05    0.05    0.06
#> prob_infected[15]         0.06    0.00 0.01    0.04    0.05    0.06    0.06
#> prob_infected[16]         0.06    0.00 0.01    0.05    0.05    0.06    0.07
#> prob_infected[17]         0.06    0.00 0.01    0.05    0.06    0.06    0.07
#> prob_infected[18]         0.07    0.00 0.01    0.05    0.06    0.07    0.07
#> prob_infected[19]         0.07    0.00 0.01    0.05    0.06    0.07    0.08
#> prob_infected[20]         0.08    0.00 0.01    0.06    0.07    0.07    0.08
#> prob_infected[21]         0.08    0.00 0.01    0.06    0.07    0.08    0.09
#> prob_infected[22]         0.08    0.00 0.01    0.06    0.07    0.08    0.09
#> prob_infected[23]         0.09    0.00 0.01    0.07    0.08    0.09    0.09
#> prob_infected[24]         0.09    0.00 0.01    0.07    0.08    0.09    0.10
#> prob_infected[25]         0.09    0.00 0.01    0.07    0.08    0.09    0.10
#> prob_infected[26]         0.10    0.00 0.01    0.07    0.09    0.10    0.10
#> prob_infected[27]         0.10    0.00 0.01    0.08    0.09    0.10    0.11
#> prob_infected[28]         0.10    0.00 0.01    0.08    0.09    0.10    0.11
#> prob_infected[29]         0.11    0.00 0.01    0.08    0.10    0.11    0.12
#> prob_infected[30]         0.11    0.00 0.01    0.08    0.10    0.11    0.12
#> prob_infected[31]         0.11    0.00 0.02    0.09    0.10    0.11    0.12
#> prob_infected[32]         0.12    0.00 0.02    0.09    0.11    0.12    0.13
#> prob_infected[33]         0.12    0.00 0.02    0.09    0.11    0.12    0.13
#> prob_infected[34]         0.12    0.00 0.02    0.09    0.11    0.12    0.13
#> prob_infected[35]         0.13    0.00 0.02    0.10    0.12    0.13    0.14
#> prob_infected[36]         0.13    0.00 0.02    0.10    0.12    0.13    0.14
#> prob_infected[37]         0.13    0.00 0.02    0.10    0.12    0.13    0.15
#> prob_infected[38]         0.14    0.00 0.02    0.11    0.13    0.14    0.15
#> prob_infected[39]         0.14    0.00 0.02    0.11    0.13    0.14    0.15
#> prob_infected[40]         0.14    0.00 0.02    0.11    0.13    0.14    0.16
#> prob_infected[41]         0.15    0.00 0.02    0.11    0.13    0.15    0.16
#> prob_infected[42]         0.15    0.00 0.02    0.12    0.14    0.15    0.16
#> prob_infected[43]         0.15    0.00 0.02    0.12    0.14    0.15    0.17
#> prob_infected[44]         0.16    0.00 0.02    0.12    0.14    0.16    0.17
#> prob_infected[45]         0.16    0.00 0.02    0.12    0.15    0.16    0.17
#> prob_infected[46]         0.16    0.00 0.02    0.13    0.15    0.16    0.18
#> prob_infected[47]         0.17    0.00 0.02    0.13    0.16    0.17    0.18
#> prob_infected[48]         0.17    0.00 0.02    0.13    0.16    0.17    0.19
#> prob_infected[49]         0.18    0.00 0.02    0.14    0.16    0.18    0.19
#> prob_infected[50]         0.18    0.00 0.02    0.14    0.16    0.18    0.20
#> prob_infected[51]         0.18    0.00 0.02    0.14    0.17    0.18    0.20
#> prob_infected[52]         0.19    0.00 0.02    0.14    0.17    0.19    0.20
#> prob_infected[53]         0.19    0.00 0.02    0.15    0.17    0.19    0.21
#> prob_infected[54]         0.19    0.00 0.02    0.15    0.18    0.19    0.21
#> prob_infected[55]         0.20    0.00 0.02    0.15    0.18    0.19    0.21
#> prob_infected[56]         0.20    0.00 0.02    0.15    0.18    0.20    0.22
#> prob_infected[57]         0.20    0.00 0.03    0.16    0.18    0.20    0.22
#> prob_infected[58]         0.21    0.00 0.03    0.16    0.19    0.20    0.22
#> prob_infected[59]         0.21    0.00 0.03    0.16    0.19    0.21    0.23
#> prob_infected[60]         0.21    0.00 0.03    0.16    0.19    0.21    0.23
#> prob_infected[61]         0.21    0.00 0.03    0.17    0.20    0.21    0.23
#> prob_infected[62]         0.22    0.00 0.03    0.17    0.20    0.22    0.24
#> prob_infected[63]         0.22    0.00 0.03    0.17    0.20    0.22    0.24
#> prob_infected[64]         0.22    0.00 0.03    0.17    0.20    0.22    0.24
#> prob_infected[65]         0.23    0.00 0.03    0.18    0.21    0.23    0.25
#> prob_infected[66]         0.23    0.00 0.03    0.18    0.21    0.23    0.25
#> prob_infected[67]         0.23    0.00 0.03    0.18    0.21    0.23    0.25
#> prob_infected[68]         0.24    0.00 0.03    0.18    0.22    0.23    0.25
#> prob_infected[69]         0.24    0.00 0.03    0.19    0.22    0.24    0.26
#> prob_infected[70]         0.24    0.00 0.03    0.19    0.22    0.24    0.26
#> prob_infected[71]         0.24    0.00 0.03    0.19    0.22    0.24    0.26
#> prob_infected[72]         0.26    0.00 0.03    0.20    0.24    0.26    0.28
#> scalar_dot_product[1]     0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> scalar_dot_product[2]     0.01    0.00 0.00    0.01    0.01    0.01    0.01
#> scalar_dot_product[3]     0.01    0.00 0.00    0.01    0.01    0.01    0.01
#> scalar_dot_product[4]     0.02    0.00 0.00    0.01    0.01    0.02    0.02
#> scalar_dot_product[5]     0.02    0.00 0.00    0.01    0.02    0.02    0.02
#> scalar_dot_product[6]     0.02    0.00 0.00    0.02    0.02    0.02    0.03
#> scalar_dot_product[7]     0.03    0.00 0.00    0.02    0.02    0.03    0.03
#> scalar_dot_product[8]     0.03    0.00 0.00    0.02    0.03    0.03    0.03
#> scalar_dot_product[9]     0.04    0.00 0.00    0.03    0.03    0.03    0.04
#> scalar_dot_product[10]    0.04    0.00 0.01    0.03    0.04    0.04    0.04
#> scalar_dot_product[11]    0.04    0.00 0.01    0.03    0.04    0.04    0.05
#> scalar_dot_product[12]    0.05    0.00 0.01    0.04    0.04    0.05    0.05
#> scalar_dot_product[13]    0.05    0.00 0.01    0.04    0.05    0.05    0.06
#> scalar_dot_product[14]    0.05    0.00 0.01    0.04    0.05    0.05    0.06
#> scalar_dot_product[15]    0.06    0.00 0.01    0.04    0.05    0.06    0.06
#> scalar_dot_product[16]    0.06    0.00 0.01    0.05    0.06    0.06    0.07
#> scalar_dot_product[17]    0.07    0.00 0.01    0.05    0.06    0.07    0.07
#> scalar_dot_product[18]    0.07    0.00 0.01    0.05    0.06    0.07    0.08
#> scalar_dot_product[19]    0.07    0.00 0.01    0.06    0.07    0.07    0.08
#> scalar_dot_product[20]    0.08    0.00 0.01    0.06    0.07    0.08    0.09
#> scalar_dot_product[21]    0.08    0.00 0.01    0.06    0.07    0.08    0.09
#> scalar_dot_product[22]    0.09    0.00 0.01    0.06    0.08    0.09    0.09
#> scalar_dot_product[23]    0.09    0.00 0.01    0.07    0.08    0.09    0.10
#> scalar_dot_product[24]    0.09    0.00 0.01    0.07    0.08    0.09    0.10
#> scalar_dot_product[25]    0.10    0.00 0.01    0.07    0.09    0.10    0.11
#> scalar_dot_product[26]    0.10    0.00 0.01    0.08    0.09    0.10    0.11
#> scalar_dot_product[27]    0.11    0.00 0.01    0.08    0.10    0.10    0.12
#> scalar_dot_product[28]    0.11    0.00 0.02    0.08    0.10    0.11    0.12
#> scalar_dot_product[29]    0.11    0.00 0.02    0.09    0.10    0.11    0.12
#> scalar_dot_product[30]    0.12    0.00 0.02    0.09    0.11    0.12    0.13
#> scalar_dot_product[31]    0.12    0.00 0.02    0.09    0.11    0.12    0.13
#> scalar_dot_product[32]    0.13    0.00 0.02    0.09    0.11    0.12    0.14
#> scalar_dot_product[33]    0.13    0.00 0.02    0.10    0.12    0.13    0.14
#> scalar_dot_product[34]    0.13    0.00 0.02    0.10    0.12    0.13    0.14
#> scalar_dot_product[35]    0.14    0.00 0.02    0.10    0.12    0.14    0.15
#> scalar_dot_product[36]    0.14    0.00 0.02    0.11    0.13    0.14    0.15
#> scalar_dot_product[37]    0.14    0.00 0.02    0.11    0.13    0.14    0.16
#> scalar_dot_product[38]    0.15    0.00 0.02    0.11    0.13    0.15    0.16
#> scalar_dot_product[39]    0.15    0.00 0.02    0.11    0.14    0.15    0.17
#> scalar_dot_product[40]    0.16    0.00 0.02    0.12    0.14    0.15    0.17
#> scalar_dot_product[41]    0.16    0.00 0.02    0.12    0.14    0.16    0.17
#> scalar_dot_product[42]    0.16    0.00 0.02    0.12    0.15    0.16    0.18
#> scalar_dot_product[43]    0.17    0.00 0.02    0.13    0.15    0.17    0.18
#> scalar_dot_product[44]    0.17    0.00 0.02    0.13    0.15    0.17    0.19
#> scalar_dot_product[45]    0.18    0.00 0.02    0.13    0.16    0.17    0.19
#> scalar_dot_product[46]    0.18    0.00 0.03    0.13    0.16    0.18    0.20
#> scalar_dot_product[47]    0.19    0.00 0.03    0.14    0.17    0.19    0.20
#> scalar_dot_product[48]    0.19    0.00 0.03    0.14    0.17    0.19    0.21
#> scalar_dot_product[49]    0.20    0.00 0.03    0.15    0.18    0.19    0.21
#> scalar_dot_product[50]    0.20    0.00 0.03    0.15    0.18    0.20    0.22
#> scalar_dot_product[51]    0.20    0.00 0.03    0.15    0.18    0.20    0.22
#> scalar_dot_product[52]    0.21    0.00 0.03    0.16    0.19    0.21    0.23
#> scalar_dot_product[53]    0.21    0.00 0.03    0.16    0.19    0.21    0.23
#> scalar_dot_product[54]    0.22    0.00 0.03    0.16    0.19    0.21    0.23
#> scalar_dot_product[55]    0.22    0.00 0.03    0.16    0.20    0.22    0.24
#> scalar_dot_product[56]    0.22    0.00 0.03    0.17    0.20    0.22    0.24
#> scalar_dot_product[57]    0.23    0.00 0.03    0.17    0.20    0.22    0.25
#> scalar_dot_product[58]    0.23    0.00 0.03    0.17    0.21    0.23    0.25
#> scalar_dot_product[59]    0.23    0.00 0.03    0.18    0.21    0.23    0.26
#> scalar_dot_product[60]    0.24    0.00 0.03    0.18    0.21    0.24    0.26
#> scalar_dot_product[61]    0.24    0.00 0.03    0.18    0.22    0.24    0.26
#> scalar_dot_product[62]    0.25    0.00 0.03    0.18    0.22    0.24    0.27
#> scalar_dot_product[63]    0.25    0.00 0.04    0.19    0.23    0.25    0.27
#> scalar_dot_product[64]    0.25    0.00 0.04    0.19    0.23    0.25    0.28
#> scalar_dot_product[65]    0.26    0.00 0.04    0.19    0.23    0.26    0.28
#> scalar_dot_product[66]    0.26    0.00 0.04    0.20    0.24    0.26    0.29
#> scalar_dot_product[67]    0.27    0.00 0.04    0.20    0.24    0.26    0.29
#> scalar_dot_product[68]    0.27    0.00 0.04    0.20    0.24    0.27    0.29
#> scalar_dot_product[69]    0.27    0.00 0.04    0.21    0.25    0.27    0.30
#> scalar_dot_product[70]    0.28    0.00 0.04    0.21    0.25    0.27    0.30
#> scalar_dot_product[71]    0.28    0.00 0.04    0.21    0.25    0.28    0.31
#> scalar_dot_product[72]    0.30    0.00 0.04    0.23    0.27    0.30    0.33
#> foi[1]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[2]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[3]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[4]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[5]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[6]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[7]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[8]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[9]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[10]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[11]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[12]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[13]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[14]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[15]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[16]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[17]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[18]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[19]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[20]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[21]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[22]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[23]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[24]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[25]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[26]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[27]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[28]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[29]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[30]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[31]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[32]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[33]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[34]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[35]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[36]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[37]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[38]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[39]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[40]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[41]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[42]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[43]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[44]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[45]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[46]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[47]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[48]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[49]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[50]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[51]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[52]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[53]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[54]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[55]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[56]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[57]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[58]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[59]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[60]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[61]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[62]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[63]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[64]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[65]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[66]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[67]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[68]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[69]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[70]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[71]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[72]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[73]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[74]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[75]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[76]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[77]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> n_pos_sim[1]              0.14    0.01 0.39    0.00    0.00    0.00    0.00
#> n_pos_sim[2]              0.21    0.01 0.46    0.00    0.00    0.00    0.00
#> n_pos_sim[3]              0.40    0.02 0.64    0.00    0.00    0.00    1.00
#> n_pos_sim[4]              0.45    0.02 0.67    0.00    0.00    0.00    1.00
#> n_pos_sim[5]              0.69    0.02 0.80    0.00    0.00    1.00    1.00
#> n_pos_sim[6]              0.52    0.02 0.71    0.00    0.00    0.00    1.00
#> n_pos_sim[7]              0.93    0.02 0.97    0.00    0.00    1.00    1.00
#> n_pos_sim[8]              0.55    0.02 0.74    0.00    0.00    0.00    1.00
#> n_pos_sim[9]              0.73    0.02 0.84    0.00    0.00    1.00    1.00
#> n_pos_sim[10]             0.93    0.02 0.97    0.00    0.00    1.00    1.00
#> n_pos_sim[11]             0.55    0.02 0.71    0.00    0.00    0.00    1.00
#> n_pos_sim[12]             1.36    0.03 1.13    0.00    1.00    1.00    2.00
#> n_pos_sim[13]             0.75    0.02 0.85    0.00    0.00    1.00    1.00
#> n_pos_sim[14]             1.33    0.03 1.15    0.00    0.00    1.00    2.00
#> n_pos_sim[15]             0.74    0.02 0.84    0.00    0.00    1.00    1.00
#> n_pos_sim[16]             0.97    0.02 0.97    0.00    0.00    1.00    2.00
#> n_pos_sim[17]             0.76    0.02 0.87    0.00    0.00    1.00    1.00
#> n_pos_sim[18]             1.20    0.02 1.09    0.00    0.00    1.00    2.00
#> n_pos_sim[19]             0.65    0.02 0.77    0.00    0.00    0.00    1.00
#> n_pos_sim[20]             0.60    0.02 0.76    0.00    0.00    0.00    1.00
#> n_pos_sim[21]             0.32    0.01 0.56    0.00    0.00    0.00    1.00
#> n_pos_sim[22]             1.55    0.03 1.21    0.00    1.00    1.00    2.00
#> n_pos_sim[23]             1.29    0.03 1.09    0.00    0.00    1.00    2.00
#> n_pos_sim[24]             1.26    0.02 1.09    0.00    0.00    1.00    2.00
#> n_pos_sim[25]             1.88    0.03 1.33    0.00    1.00    2.00    3.00
#> n_pos_sim[26]             1.07    0.02 0.99    0.00    0.00    1.00    2.00
#> n_pos_sim[27]             0.79    0.02 0.86    0.00    0.00    1.00    1.00
#> n_pos_sim[28]             0.91    0.02 0.90    0.00    0.00    1.00    1.00
#> n_pos_sim[29]             0.93    0.02 0.91    0.00    0.00    1.00    1.00
#> n_pos_sim[30]             1.11    0.02 1.01    0.00    0.00    1.00    2.00
#> n_pos_sim[31]             1.16    0.02 1.03    0.00    0.00    1.00    2.00
#> n_pos_sim[32]             1.24    0.02 1.05    0.00    0.00    1.00    2.00
#> n_pos_sim[33]             1.35    0.03 1.10    0.00    1.00    1.00    2.00
#> n_pos_sim[34]             0.99    0.02 0.96    0.00    0.00    1.00    2.00
#> n_pos_sim[35]             1.32    0.02 1.07    0.00    1.00    1.00    2.00
#> n_pos_sim[36]             1.37    0.03 1.07    0.00    1.00    1.00    2.00
#> n_pos_sim[37]             1.24    0.02 1.02    0.00    0.00    1.00    2.00
#> n_pos_sim[38]             0.55    0.02 0.69    0.00    0.00    0.00    1.00
#> n_pos_sim[39]             0.55    0.02 0.69    0.00    0.00    0.00    1.00
#> n_pos_sim[40]             1.29    0.02 1.05    0.00    1.00    1.00    2.00
#> n_pos_sim[41]             0.60    0.02 0.72    0.00    0.00    0.00    1.00
#> n_pos_sim[42]             1.21    0.02 1.02    0.00    0.00    1.00    2.00
#> n_pos_sim[43]             0.49    0.02 0.65    0.00    0.00    0.00    1.00
#> n_pos_sim[44]             0.16    0.01 0.37    0.00    0.00    0.00    0.00
#> n_pos_sim[45]             0.49    0.01 0.65    0.00    0.00    0.00    1.00
#> n_pos_sim[46]             0.50    0.02 0.65    0.00    0.00    0.00    1.00
#> n_pos_sim[47]             0.34    0.01 0.53    0.00    0.00    0.00    1.00
#> n_pos_sim[48]             0.53    0.02 0.66    0.00    0.00    0.00    1.00
#> n_pos_sim[49]             0.51    0.02 0.66    0.00    0.00    0.00    1.00
#> n_pos_sim[50]             0.19    0.01 0.39    0.00    0.00    0.00    0.00
#> n_pos_sim[51]             1.11    0.02 0.96    0.00    0.00    1.00    2.00
#> n_pos_sim[52]             0.75    0.02 0.78    0.00    0.00    1.00    1.00
#> n_pos_sim[53]             0.58    0.02 0.68    0.00    0.00    0.00    1.00
#> n_pos_sim[54]             1.13    0.02 0.95    0.00    0.00    1.00    2.00
#> n_pos_sim[55]             0.19    0.01 0.39    0.00    0.00    0.00    0.00
#> n_pos_sim[56]             0.76    0.02 0.78    0.00    0.00    1.00    1.00
#> n_pos_sim[57]             0.61    0.02 0.70    0.00    0.00    0.00    1.00
#> n_pos_sim[58]             0.21    0.01 0.41    0.00    0.00    0.00    0.00
#> n_pos_sim[59]             2.27    0.03 1.39    0.00    1.00    2.00    3.00
#> n_pos_sim[60]             1.07    0.02 0.94    0.00    0.00    1.00    2.00
#> n_pos_sim[61]             0.21    0.01 0.41    0.00    0.00    0.00    0.00
#> n_pos_sim[62]             0.43    0.01 0.58    0.00    0.00    0.00    1.00
#> n_pos_sim[63]             1.11    0.02 0.94    0.00    0.00    1.00    2.00
#> n_pos_sim[64]             1.79    0.03 1.16    0.00    1.00    2.00    3.00
#> n_pos_sim[65]             0.23    0.01 0.42    0.00    0.00    0.00    0.00
#> n_pos_sim[66]             0.22    0.01 0.41    0.00    0.00    0.00    0.00
#> n_pos_sim[67]             0.46    0.01 0.59    0.00    0.00    0.00    1.00
#> n_pos_sim[68]             1.18    0.02 0.94    0.00    0.00    1.00    2.00
#> n_pos_sim[69]             0.71    0.02 0.73    0.00    0.00    1.00    1.00
#> n_pos_sim[70]             0.26    0.01 0.44    0.00    0.00    0.00    1.00
#> n_pos_sim[71]             0.23    0.01 0.42    0.00    0.00    0.00    0.00
#> n_pos_sim[72]             0.27    0.01 0.44    0.00    0.00    0.00    1.00
#> P_sim[1]                  0.00    0.00 0.01    0.00    0.00    0.00    0.00
#> P_sim[2]                  0.01    0.00 0.02    0.00    0.00    0.00    0.00
#> P_sim[3]                  0.01    0.00 0.02    0.00    0.00    0.00    0.03
#> P_sim[4]                  0.02    0.00 0.02    0.00    0.00    0.00    0.03
#> P_sim[5]                  0.02    0.00 0.02    0.00    0.00    0.03    0.03
#> P_sim[6]                  0.02    0.00 0.03    0.00    0.00    0.00    0.04
#> P_sim[7]                  0.03    0.00 0.03    0.00    0.00    0.03    0.03
#> P_sim[8]                  0.03    0.00 0.04    0.00    0.00    0.00    0.06
#> P_sim[9]                  0.03    0.00 0.04    0.00    0.00    0.05    0.05
#> P_sim[10]                 0.04    0.00 0.04    0.00    0.00    0.04    0.04
#> P_sim[11]                 0.04    0.00 0.05    0.00    0.00    0.00    0.08
#> P_sim[12]                 0.05    0.00 0.04    0.00    0.04    0.04    0.07
#> P_sim[13]                 0.05    0.00 0.06    0.00    0.00    0.07    0.07
#> P_sim[14]                 0.05    0.00 0.05    0.00    0.00    0.04    0.08
#> P_sim[15]                 0.06    0.00 0.06    0.00    0.00    0.08    0.08
#> P_sim[16]                 0.06    0.00 0.06    0.00    0.00    0.06    0.12
#> P_sim[17]                 0.06    0.00 0.07    0.00    0.00    0.08    0.08
#> P_sim[18]                 0.07    0.00 0.06    0.00    0.00    0.06    0.11
#> P_sim[19]                 0.07    0.00 0.09    0.00    0.00    0.00    0.11
#> P_sim[20]                 0.08    0.00 0.09    0.00    0.00    0.00    0.12
#> P_sim[21]                 0.08    0.00 0.14    0.00    0.00    0.00    0.25
#> P_sim[22]                 0.08    0.00 0.06    0.00    0.05    0.05    0.11
#> P_sim[23]                 0.09    0.00 0.07    0.00    0.00    0.07    0.13
#> P_sim[24]                 0.09    0.00 0.08    0.00    0.00    0.07    0.14
#> P_sim[25]                 0.09    0.00 0.07    0.00    0.05    0.10    0.15
#> P_sim[26]                 0.10    0.00 0.09    0.00    0.00    0.09    0.18
#> P_sim[27]                 0.10    0.00 0.11    0.00    0.00    0.12    0.12
#> P_sim[28]                 0.10    0.00 0.10    0.00    0.00    0.11    0.11
#> P_sim[29]                 0.10    0.00 0.10    0.00    0.00    0.11    0.11
#> P_sim[30]                 0.11    0.00 0.10    0.00    0.00    0.10    0.20
#> P_sim[31]                 0.12    0.00 0.10    0.00    0.00    0.10    0.20
#> P_sim[32]                 0.11    0.00 0.10    0.00    0.00    0.09    0.18
#> P_sim[33]                 0.12    0.00 0.10    0.00    0.09    0.09    0.18
#> P_sim[34]                 0.12    0.00 0.12    0.00    0.00    0.12    0.25
#> P_sim[35]                 0.13    0.00 0.11    0.00    0.10    0.10    0.20
#> P_sim[36]                 0.14    0.00 0.11    0.00    0.10    0.10    0.20
#> P_sim[37]                 0.14    0.00 0.11    0.00    0.00    0.11    0.22
#> P_sim[38]                 0.14    0.00 0.17    0.00    0.00    0.00    0.25
#> P_sim[39]                 0.14    0.00 0.17    0.00    0.00    0.00    0.25
#> P_sim[40]                 0.14    0.00 0.12    0.00    0.11    0.11    0.22
#> P_sim[41]                 0.15    0.00 0.18    0.00    0.00    0.00    0.25
#> P_sim[42]                 0.15    0.00 0.13    0.00    0.00    0.12    0.25
#> P_sim[43]                 0.16    0.01 0.22    0.00    0.00    0.00    0.33
#> P_sim[44]                 0.16    0.01 0.37    0.00    0.00    0.00    0.00
#> P_sim[45]                 0.16    0.00 0.22    0.00    0.00    0.00    0.33
#> P_sim[46]                 0.17    0.01 0.22    0.00    0.00    0.00    0.33
#> P_sim[47]                 0.17    0.01 0.26    0.00    0.00    0.00    0.50
#> P_sim[48]                 0.18    0.01 0.22    0.00    0.00    0.00    0.33
#> P_sim[49]                 0.17    0.01 0.22    0.00    0.00    0.00    0.33
#> P_sim[50]                 0.19    0.01 0.39    0.00    0.00    0.00    0.00
#> P_sim[51]                 0.18    0.00 0.16    0.00    0.00    0.17    0.33
#> P_sim[52]                 0.19    0.00 0.20    0.00    0.00    0.25    0.25
#> P_sim[53]                 0.19    0.01 0.23    0.00    0.00    0.00    0.33
#> P_sim[54]                 0.19    0.00 0.16    0.00    0.00    0.17    0.33
#> P_sim[55]                 0.19    0.01 0.39    0.00    0.00    0.00    0.00
#> P_sim[56]                 0.19    0.00 0.20    0.00    0.00    0.25    0.25
#> P_sim[57]                 0.20    0.01 0.23    0.00    0.00    0.00    0.33
#> P_sim[58]                 0.21    0.01 0.41    0.00    0.00    0.00    0.00
#> P_sim[59]                 0.21    0.00 0.13    0.00    0.09    0.18    0.27
#> P_sim[60]                 0.22    0.00 0.19    0.00    0.00    0.20    0.40
#> P_sim[61]                 0.21    0.01 0.41    0.00    0.00    0.00    0.00
#> P_sim[62]                 0.22    0.01 0.29    0.00    0.00    0.00    0.50
#> P_sim[63]                 0.22    0.00 0.19    0.00    0.00    0.20    0.40
#> P_sim[64]                 0.22    0.00 0.14    0.00    0.12    0.25    0.38
#> P_sim[65]                 0.23    0.01 0.42    0.00    0.00    0.00    0.00
#> P_sim[66]                 0.22    0.01 0.41    0.00    0.00    0.00    0.00
#> P_sim[67]                 0.23    0.01 0.30    0.00    0.00    0.00    0.50
#> P_sim[68]                 0.24    0.00 0.19    0.00    0.00    0.20    0.40
#> P_sim[69]                 0.24    0.01 0.24    0.00    0.00    0.33    0.33
#> P_sim[70]                 0.26    0.01 0.44    0.00    0.00    0.00    1.00
#> P_sim[71]                 0.23    0.01 0.42    0.00    0.00    0.00    0.00
#> P_sim[72]                 0.27    0.01 0.44    0.00    0.00    0.00    1.00
#> logLikelihood[1]         -0.13    0.00 0.02   -0.17   -0.14   -0.13   -0.12
#> logLikelihood[2]         -0.20    0.00 0.03   -0.25   -0.21   -0.19   -0.18
#> logLikelihood[3]         -1.30    0.00 0.08   -1.48   -1.36   -1.30   -1.25
#> logLikelihood[4]         -0.45    0.00 0.06   -0.59   -0.49   -0.45   -0.41
#> logLikelihood[5]         -0.70    0.00 0.10   -0.92   -0.77   -0.70   -0.63
#> logLikelihood[6]         -0.54    0.00 0.08   -0.70   -0.59   -0.53   -0.49
#> logLikelihood[7]         -0.93    0.00 0.13   -1.21   -1.01   -0.92   -0.84
#> logLikelihood[8]         -0.56    0.00 0.08   -0.73   -0.61   -0.56   -0.51
#> logLikelihood[9]         -0.74    0.00 0.10   -0.96   -0.81   -0.73   -0.67
#> logLikelihood[10]        -0.94    0.00 0.13   -1.22   -1.02   -0.93   -0.85
#> logLikelihood[11]        -1.13    0.00 0.07   -1.27   -1.17   -1.12   -1.08
#> logLikelihood[12]        -1.31    0.01 0.18   -1.71   -1.43   -1.30   -1.18
#> logLikelihood[13]        -0.76    0.00 0.11   -0.99   -0.83   -0.75   -0.69
#> logLikelihood[14]        -1.37    0.01 0.19   -1.78   -1.49   -1.35   -1.23
#> logLikelihood[15]        -0.76    0.00 0.11   -0.99   -0.83   -0.75   -0.69
#> logLikelihood[16]        -1.00    0.01 0.14   -1.30   -1.09   -0.99   -0.90
#> logLikelihood[17]        -0.80    0.00 0.11   -1.04   -0.87   -0.79   -0.72
#> logLikelihood[18]        -1.27    0.01 0.18   -1.65   -1.38   -1.25   -1.14
#> logLikelihood[19]        -0.67    0.00 0.09   -0.87   -0.73   -0.66   -0.60
#> logLikelihood[20]        -0.63    0.00 0.09   -0.82   -0.68   -0.62   -0.56
#> logLikelihood[21]        -0.33    0.00 0.05   -0.43   -0.36   -0.33   -0.30
#> logLikelihood[22]        -1.63    0.01 0.23   -2.13   -1.78   -1.62   -1.47
#> logLikelihood[23]        -1.35    0.01 0.19   -1.76   -1.47   -1.34   -1.21
#> logLikelihood[24]        -1.31    0.01 0.18   -1.71   -1.43   -1.30   -1.18
#> logLikelihood[25]        -1.95    0.01 0.27   -2.55   -2.13   -1.94   -1.76
#> logLikelihood[26]        -0.97    0.00 0.02   -1.01   -0.97   -0.96   -0.95
#> logLikelihood[27]        -0.97    0.00 0.03   -1.05   -0.99   -0.96   -0.95
#> logLikelihood[28]        -0.99    0.00 0.14   -1.28   -1.07   -0.98   -0.89
#> logLikelihood[29]        -1.02    0.01 0.14   -1.33   -1.11   -1.01   -0.92
#> logLikelihood[30]        -0.96    0.00 0.02   -1.03   -0.97   -0.96   -0.95
#> logLikelihood[31]        -0.97    0.00 0.03   -1.04   -0.98   -0.96   -0.95
#> logLikelihood[32]        -1.38    0.01 0.19   -1.79   -1.50   -1.36   -1.24
#> logLikelihood[33]        -1.01    0.00 0.05   -1.15   -1.04   -1.00   -0.97
#> logLikelihood[34]        -1.06    0.01 0.15   -1.39   -1.16   -1.05   -0.96
#> logLikelihood[35]        -1.00    0.00 0.04   -1.11   -1.02   -0.98   -0.96
#> logLikelihood[36]        -2.32    0.01 0.25   -2.85   -2.48   -2.31   -2.14
#> logLikelihood[37]        -0.97    0.00 0.04   -1.07   -0.99   -0.96   -0.95
#> logLikelihood[38]        -1.05    0.00 0.07   -1.20   -1.09   -1.04   -1.00
#> logLikelihood[39]        -0.61    0.00 0.09   -0.79   -0.66   -0.60   -0.55
#> logLikelihood[40]        -1.40    0.00 0.11   -1.64   -1.46   -1.38   -1.32
#> logLikelihood[41]        -2.37    0.01 0.21   -2.80   -2.51   -2.36   -2.22
#> logLikelihood[42]        -1.31    0.01 0.18   -1.71   -1.43   -1.30   -1.18
#> logLikelihood[43]        -1.11    0.00 0.08   -1.29   -1.17   -1.11   -1.06
#> logLikelihood[44]        -1.85    0.00 0.13   -2.11   -1.94   -1.85   -1.77
#> logLikelihood[45]        -0.53    0.00 0.07   -0.69   -0.58   -0.52   -0.48
#> logLikelihood[46]        -2.71    0.01 0.23   -3.18   -2.86   -2.71   -2.55
#> logLikelihood[47]        -3.55    0.01 0.26   -4.06   -3.72   -3.55   -3.38
#> logLikelihood[48]        -2.61    0.01 0.23   -3.07   -2.76   -2.60   -2.45
#> logLikelihood[49]        -0.59    0.00 0.08   -0.76   -0.64   -0.58   -0.53
#> logLikelihood[50]        -1.72    0.00 0.13   -1.97   -1.81   -1.72   -1.63
#> logLikelihood[51]        -2.72    0.01 0.29   -3.33   -2.92   -2.71   -2.52
#> logLikelihood[52]        -0.83    0.00 0.12   -1.08   -0.90   -0.82   -0.75
#> logLikelihood[53]        -0.99    0.00 0.07   -1.14   -1.04   -0.99   -0.94
#> logLikelihood[54]        -2.61    0.01 0.29   -3.20   -2.80   -2.60   -2.40
#> logLikelihood[55]        -0.22    0.00 0.03   -0.29   -0.24   -0.22   -0.20
#> logLikelihood[56]        -3.70    0.01 0.34   -4.39   -3.93   -3.69   -3.46
#> logLikelihood[57]        -2.34    0.01 0.22   -2.78   -2.48   -2.33   -2.19
#> logLikelihood[58]        -0.23    0.00 0.03   -0.30   -0.25   -0.23   -0.21
#> logLikelihood[59]        -1.52    0.01 0.21   -1.99   -1.65   -1.50   -1.37
#> logLikelihood[60]        -1.19    0.01 0.17   -1.55   -1.30   -1.18   -1.07
#> logLikelihood[61]        -1.55    0.00 0.12   -1.79   -1.63   -1.54   -1.46
#> logLikelihood[62]        -1.08    0.00 0.09   -1.27   -1.14   -1.08   -1.02
#> logLikelihood[63]        -2.75    0.01 0.30   -3.37   -2.95   -2.74   -2.54
#> logLikelihood[64]        -1.20    0.00 0.04   -1.31   -1.22   -1.19   -1.17
#> logLikelihood[65]        -1.49    0.00 0.12   -1.74   -1.57   -1.49   -1.41
#> logLikelihood[66]        -1.48    0.00 0.12   -1.72   -1.56   -1.48   -1.39
#> logLikelihood[67]        -0.53    0.00 0.07   -0.69   -0.58   -0.53   -0.48
#> logLikelihood[68]        -1.41    0.00 0.13   -1.70   -1.49   -1.40   -1.31
#> logLikelihood[69]        -2.05    0.01 0.21   -2.47   -2.19   -2.05   -1.91
#> logLikelihood[70]        -1.43    0.00 0.12   -1.67   -1.51   -1.43   -1.34
#> logLikelihood[71]        -1.41    0.00 0.12   -1.66   -1.50   -1.41   -1.33
#> logLikelihood[72]        -1.36    0.00 0.12   -1.60   -1.44   -1.36   -1.27
#> lp__                   -156.09    0.04 0.77 -158.34 -156.27 -155.80 -155.61
#>                          97.5% n_eff Rhat
#> lambda0                   0.01   777 1.00
#> prob_infected[1]          0.01   777 1.00
#> prob_infected[2]          0.01   777 1.00
#> prob_infected[3]          0.02   777 1.00
#> prob_infected[4]          0.02   777 1.00
#> prob_infected[5]          0.03   777 1.00
#> prob_infected[6]          0.03   777 1.00
#> prob_infected[7]          0.04   777 1.00
#> prob_infected[8]          0.04   777 1.00
#> prob_infected[9]          0.04   777 1.00
#> prob_infected[10]         0.05   777 1.00
#> prob_infected[11]         0.05   777 1.00
#> prob_infected[12]         0.06   777 1.00
#> prob_infected[13]         0.06   777 1.00
#> prob_infected[14]         0.07   777 1.00
#> prob_infected[15]         0.07   777 1.00
#> prob_infected[16]         0.08   777 1.00
#> prob_infected[17]         0.08   777 1.00
#> prob_infected[18]         0.09   777 1.00
#> prob_infected[19]         0.09   777 1.00
#> prob_infected[20]         0.10   777 1.00
#> prob_infected[21]         0.10   777 1.00
#> prob_infected[22]         0.11   777 1.00
#> prob_infected[23]         0.11   777 1.00
#> prob_infected[24]         0.12   777 1.00
#> prob_infected[25]         0.12   777 1.00
#> prob_infected[26]         0.12   777 1.00
#> prob_infected[27]         0.13   777 1.00
#> prob_infected[28]         0.13   777 1.00
#> prob_infected[29]         0.14   777 1.00
#> prob_infected[30]         0.14   777 1.00
#> prob_infected[31]         0.15   777 1.00
#> prob_infected[32]         0.15   777 1.00
#> prob_infected[33]         0.15   777 1.00
#> prob_infected[34]         0.16   777 1.00
#> prob_infected[35]         0.16   777 1.00
#> prob_infected[36]         0.17   777 1.00
#> prob_infected[37]         0.17   777 1.00
#> prob_infected[38]         0.18   777 1.00
#> prob_infected[39]         0.18   777 1.00
#> prob_infected[40]         0.18   777 1.00
#> prob_infected[41]         0.19   777 1.00
#> prob_infected[42]         0.19   777 1.00
#> prob_infected[43]         0.20   777 1.00
#> prob_infected[44]         0.20   777 1.00
#> prob_infected[45]         0.20   777 1.00
#> prob_infected[46]         0.21   777 1.00
#> prob_infected[47]         0.22   777 1.00
#> prob_infected[48]         0.22   777 1.00
#> prob_infected[49]         0.22   777 1.00
#> prob_infected[50]         0.23   777 1.00
#> prob_infected[51]         0.23   777 1.00
#> prob_infected[52]         0.24   777 1.00
#> prob_infected[53]         0.24   777 1.00
#> prob_infected[54]         0.24   777 1.00
#> prob_infected[55]         0.25   777 1.00
#> prob_infected[56]         0.25   777 1.00
#> prob_infected[57]         0.26   777 1.00
#> prob_infected[58]         0.26   777 1.00
#> prob_infected[59]         0.26   777 1.00
#> prob_infected[60]         0.27   777 1.00
#> prob_infected[61]         0.27   777 1.00
#> prob_infected[62]         0.27   777 1.00
#> prob_infected[63]         0.28   777 1.00
#> prob_infected[64]         0.28   777 1.00
#> prob_infected[65]         0.29   777 1.00
#> prob_infected[66]         0.29   777 1.00
#> prob_infected[67]         0.29   777 1.00
#> prob_infected[68]         0.30   777 1.00
#> prob_infected[69]         0.30   777 1.00
#> prob_infected[70]         0.30   777 1.00
#> prob_infected[71]         0.31   777 1.00
#> prob_infected[72]         0.32   777 1.00
#> scalar_dot_product[1]     0.01   777 1.00
#> scalar_dot_product[2]     0.01   777 1.00
#> scalar_dot_product[3]     0.02   777 1.00
#> scalar_dot_product[4]     0.02   777 1.00
#> scalar_dot_product[5]     0.03   777 1.00
#> scalar_dot_product[6]     0.03   777 1.00
#> scalar_dot_product[7]     0.04   777 1.00
#> scalar_dot_product[8]     0.04   777 1.00
#> scalar_dot_product[9]     0.05   777 1.00
#> scalar_dot_product[10]    0.05   777 1.00
#> scalar_dot_product[11]    0.06   777 1.00
#> scalar_dot_product[12]    0.06   777 1.00
#> scalar_dot_product[13]    0.07   777 1.00
#> scalar_dot_product[14]    0.07   777 1.00
#> scalar_dot_product[15]    0.08   777 1.00
#> scalar_dot_product[16]    0.08   777 1.00
#> scalar_dot_product[17]    0.09   777 1.00
#> scalar_dot_product[18]    0.09   777 1.00
#> scalar_dot_product[19]    0.10   777 1.00
#> scalar_dot_product[20]    0.10   777 1.00
#> scalar_dot_product[21]    0.11   777 1.00
#> scalar_dot_product[22]    0.11   777 1.00
#> scalar_dot_product[23]    0.12   777 1.00
#> scalar_dot_product[24]    0.12   777 1.00
#> scalar_dot_product[25]    0.13   777 1.00
#> scalar_dot_product[26]    0.13   777 1.00
#> scalar_dot_product[27]    0.14   777 1.00
#> scalar_dot_product[28]    0.14   777 1.00
#> scalar_dot_product[29]    0.15   777 1.00
#> scalar_dot_product[30]    0.15   777 1.00
#> scalar_dot_product[31]    0.16   777 1.00
#> scalar_dot_product[32]    0.16   777 1.00
#> scalar_dot_product[33]    0.17   777 1.00
#> scalar_dot_product[34]    0.17   777 1.00
#> scalar_dot_product[35]    0.18   777 1.00
#> scalar_dot_product[36]    0.18   777 1.00
#> scalar_dot_product[37]    0.19   777 1.00
#> scalar_dot_product[38]    0.19   777 1.00
#> scalar_dot_product[39]    0.20   777 1.00
#> scalar_dot_product[40]    0.20   777 1.00
#> scalar_dot_product[41]    0.21   777 1.00
#> scalar_dot_product[42]    0.21   777 1.00
#> scalar_dot_product[43]    0.22   777 1.00
#> scalar_dot_product[44]    0.22   777 1.00
#> scalar_dot_product[45]    0.23   777 1.00
#> scalar_dot_product[46]    0.23   777 1.00
#> scalar_dot_product[47]    0.24   777 1.00
#> scalar_dot_product[48]    0.25   777 1.00
#> scalar_dot_product[49]    0.25   777 1.00
#> scalar_dot_product[50]    0.26   777 1.00
#> scalar_dot_product[51]    0.26   777 1.00
#> scalar_dot_product[52]    0.27   777 1.00
#> scalar_dot_product[53]    0.28   777 1.00
#> scalar_dot_product[54]    0.28   777 1.00
#> scalar_dot_product[55]    0.29   777 1.00
#> scalar_dot_product[56]    0.29   777 1.00
#> scalar_dot_product[57]    0.30   777 1.00
#> scalar_dot_product[58]    0.30   777 1.00
#> scalar_dot_product[59]    0.31   777 1.00
#> scalar_dot_product[60]    0.31   777 1.00
#> scalar_dot_product[61]    0.32   777 1.00
#> scalar_dot_product[62]    0.32   777 1.00
#> scalar_dot_product[63]    0.33   777 1.00
#> scalar_dot_product[64]    0.33   777 1.00
#> scalar_dot_product[65]    0.34   777 1.00
#> scalar_dot_product[66]    0.34   777 1.00
#> scalar_dot_product[67]    0.35   777 1.00
#> scalar_dot_product[68]    0.35   777 1.00
#> scalar_dot_product[69]    0.36   777 1.00
#> scalar_dot_product[70]    0.36   777 1.00
#> scalar_dot_product[71]    0.37   777 1.00
#> scalar_dot_product[72]    0.39   777 1.00
#> foi[1]                    0.01   777 1.00
#> foi[2]                    0.01   777 1.00
#> foi[3]                    0.01   777 1.00
#> foi[4]                    0.01   777 1.00
#> foi[5]                    0.01   777 1.00
#> foi[6]                    0.01   777 1.00
#> foi[7]                    0.01   777 1.00
#> foi[8]                    0.01   777 1.00
#> foi[9]                    0.01   777 1.00
#> foi[10]                   0.01   777 1.00
#> foi[11]                   0.01   777 1.00
#> foi[12]                   0.01   777 1.00
#> foi[13]                   0.01   777 1.00
#> foi[14]                   0.01   777 1.00
#> foi[15]                   0.01   777 1.00
#> foi[16]                   0.01   777 1.00
#> foi[17]                   0.01   777 1.00
#> foi[18]                   0.01   777 1.00
#> foi[19]                   0.01   777 1.00
#> foi[20]                   0.01   777 1.00
#> foi[21]                   0.01   777 1.00
#> foi[22]                   0.01   777 1.00
#> foi[23]                   0.01   777 1.00
#> foi[24]                   0.01   777 1.00
#> foi[25]                   0.01   777 1.00
#> foi[26]                   0.01   777 1.00
#> foi[27]                   0.01   777 1.00
#> foi[28]                   0.01   777 1.00
#> foi[29]                   0.01   777 1.00
#> foi[30]                   0.01   777 1.00
#> foi[31]                   0.01   777 1.00
#> foi[32]                   0.01   777 1.00
#> foi[33]                   0.01   777 1.00
#> foi[34]                   0.01   777 1.00
#> foi[35]                   0.01   777 1.00
#> foi[36]                   0.01   777 1.00
#> foi[37]                   0.01   777 1.00
#> foi[38]                   0.01   777 1.00
#> foi[39]                   0.01   777 1.00
#> foi[40]                   0.01   777 1.00
#> foi[41]                   0.01   777 1.00
#> foi[42]                   0.01   777 1.00
#> foi[43]                   0.01   777 1.00
#> foi[44]                   0.01   777 1.00
#> foi[45]                   0.01   777 1.00
#> foi[46]                   0.01   777 1.00
#> foi[47]                   0.01   777 1.00
#> foi[48]                   0.01   777 1.00
#> foi[49]                   0.01   777 1.00
#> foi[50]                   0.01   777 1.00
#> foi[51]                   0.01   777 1.00
#> foi[52]                   0.01   777 1.00
#> foi[53]                   0.01   777 1.00
#> foi[54]                   0.01   777 1.00
#> foi[55]                   0.01   777 1.00
#> foi[56]                   0.01   777 1.00
#> foi[57]                   0.01   777 1.00
#> foi[58]                   0.01   777 1.00
#> foi[59]                   0.01   777 1.00
#> foi[60]                   0.01   777 1.00
#> foi[61]                   0.01   777 1.00
#> foi[62]                   0.01   777 1.00
#> foi[63]                   0.01   777 1.00
#> foi[64]                   0.01   777 1.00
#> foi[65]                   0.01   777 1.00
#> foi[66]                   0.01   777 1.00
#> foi[67]                   0.01   777 1.00
#> foi[68]                   0.01   777 1.00
#> foi[69]                   0.01   777 1.00
#> foi[70]                   0.01   777 1.00
#> foi[71]                   0.01   777 1.00
#> foi[72]                   0.01   777 1.00
#> foi[73]                   0.01   777 1.00
#> foi[74]                   0.01   777 1.00
#> foi[75]                   0.01   777 1.00
#> foi[76]                   0.01   777 1.00
#> foi[77]                   0.01   777 1.00
#> n_pos_sim[1]              1.00  1977 1.00
#> n_pos_sim[2]              1.00  2010 1.00
#> n_pos_sim[3]              2.00  1775 1.00
#> n_pos_sim[4]              2.00  1946 1.00
#> n_pos_sim[5]              3.00  2082 1.00
#> n_pos_sim[6]              2.00  2144 1.00
#> n_pos_sim[7]              3.00  1763 1.00
#> n_pos_sim[8]              2.00  1896 1.00
#> n_pos_sim[9]              3.00  1640 1.00
#> n_pos_sim[10]             3.00  1749 1.00
#> n_pos_sim[11]             2.00  1864 1.00
#> n_pos_sim[12]             4.00  1984 1.00
#> n_pos_sim[13]             3.00  2027 1.00
#> n_pos_sim[14]             4.00  1920 1.00
#> n_pos_sim[15]             3.00  1945 1.00
#> n_pos_sim[16]             3.00  1794 1.00
#> n_pos_sim[17]             3.00  1874 1.00
#> n_pos_sim[18]             4.00  1984 1.00
#> n_pos_sim[19]             3.00  1802 1.00
#> n_pos_sim[20]             2.00  1809 1.00
#> n_pos_sim[21]             2.00  2030 1.00
#> n_pos_sim[22]             4.00  1876 1.00
#> n_pos_sim[23]             4.00  1766 1.00
#> n_pos_sim[24]             4.00  1898 1.00
#> n_pos_sim[25]             5.00  1850 1.00
#> n_pos_sim[26]             3.00  1790 1.00
#> n_pos_sim[27]             3.00  1930 1.00
#> n_pos_sim[28]             3.00  1821 1.00
#> n_pos_sim[29]             3.00  1963 1.00
#> n_pos_sim[30]             3.00  1760 1.00
#> n_pos_sim[31]             3.00  1805 1.00
#> n_pos_sim[32]             4.00  1908 1.00
#> n_pos_sim[33]             4.00  1784 1.00
#> n_pos_sim[34]             3.00  1851 1.00
#> n_pos_sim[35]             4.00  1958 1.00
#> n_pos_sim[36]             4.00  1650 1.00
#> n_pos_sim[37]             3.00  1826 1.00
#> n_pos_sim[38]             2.00  1846 1.00
#> n_pos_sim[39]             2.00  2023 1.00
#> n_pos_sim[40]             4.00  1845 1.00
#> n_pos_sim[41]             2.00  2056 1.00
#> n_pos_sim[42]             3.00  1900 1.00
#> n_pos_sim[43]             2.00  1739 1.00
#> n_pos_sim[44]             1.00  1938 1.00
#> n_pos_sim[45]             2.00  2444 1.00
#> n_pos_sim[46]             2.00  1763 1.00
#> n_pos_sim[47]             2.00  2107 1.00
#> n_pos_sim[48]             2.00  1590 1.00
#> n_pos_sim[49]             2.00  1728 1.00
#> n_pos_sim[50]             1.00  2205 1.00
#> n_pos_sim[51]             3.00  1899 1.00
#> n_pos_sim[52]             2.00  1925 1.00
#> n_pos_sim[53]             2.00  1875 1.00
#> n_pos_sim[54]             3.00  1869 1.00
#> n_pos_sim[55]             1.00  1988 1.00
#> n_pos_sim[56]             3.00  1997 1.00
#> n_pos_sim[57]             2.00  2091 1.00
#> n_pos_sim[58]             1.00  1960 1.00
#> n_pos_sim[59]             5.00  1765 1.00
#> n_pos_sim[60]             3.00  2106 1.00
#> n_pos_sim[61]             1.00  1598 1.00
#> n_pos_sim[62]             2.00  1760 1.00
#> n_pos_sim[63]             3.00  1797 1.00
#> n_pos_sim[64]             4.00  2004 1.00
#> n_pos_sim[65]             1.00  2024 1.00
#> n_pos_sim[66]             1.00  2003 1.00
#> n_pos_sim[67]             2.00  1769 1.00
#> n_pos_sim[68]             3.00  1960 1.00
#> n_pos_sim[69]             2.00  1983 1.00
#> n_pos_sim[70]             1.00  1988 1.00
#> n_pos_sim[71]             1.00  1917 1.00
#> n_pos_sim[72]             1.00  1711 1.00
#> P_sim[1]                  0.03  1977 1.00
#> P_sim[2]                  0.04  2010 1.00
#> P_sim[3]                  0.06  1775 1.00
#> P_sim[4]                  0.07  1946 1.00
#> P_sim[5]                  0.08  2082 1.00
#> P_sim[6]                  0.09  2144 1.00
#> P_sim[7]                  0.09  1763 1.00
#> P_sim[8]                  0.11  1896 1.00
#> P_sim[9]                  0.14  1640 1.00
#> P_sim[10]                 0.12  1749 1.00
#> P_sim[11]                 0.15  1864 1.00
#> P_sim[12]                 0.14  1984 1.00
#> P_sim[13]                 0.20  2027 1.00
#> P_sim[14]                 0.16  1920 1.00
#> P_sim[15]                 0.23  1945 1.00
#> P_sim[16]                 0.19  1794 1.00
#> P_sim[17]                 0.25  1874 1.00
#> P_sim[18]                 0.22  1984 1.00
#> P_sim[19]                 0.33  1802 1.00
#> P_sim[20]                 0.25  1809 1.00
#> P_sim[21]                 0.50  2030 1.00
#> P_sim[22]                 0.21  1876 1.00
#> P_sim[23]                 0.27  1766 1.00
#> P_sim[24]                 0.29  1898 1.00
#> P_sim[25]                 0.25  1850 1.00
#> P_sim[26]                 0.27  1790 1.00
#> P_sim[27]                 0.38  1930 1.00
#> P_sim[28]                 0.33  1821 1.00
#> P_sim[29]                 0.33  1963 1.00
#> P_sim[30]                 0.30  1760 1.00
#> P_sim[31]                 0.30  1805 1.00
#> P_sim[32]                 0.36  1908 1.00
#> P_sim[33]                 0.36  1784 1.00
#> P_sim[34]                 0.38  1851 1.00
#> P_sim[35]                 0.40  1958 1.00
#> P_sim[36]                 0.40  1650 1.00
#> P_sim[37]                 0.33  1826 1.00
#> P_sim[38]                 0.50  1846 1.00
#> P_sim[39]                 0.50  2023 1.00
#> P_sim[40]                 0.44  1845 1.00
#> P_sim[41]                 0.50  2056 1.00
#> P_sim[42]                 0.38  1900 1.00
#> P_sim[43]                 0.67  1739 1.00
#> P_sim[44]                 1.00  1938 1.00
#> P_sim[45]                 0.67  2444 1.00
#> P_sim[46]                 0.67  1763 1.00
#> P_sim[47]                 1.00  2107 1.00
#> P_sim[48]                 0.67  1590 1.00
#> P_sim[49]                 0.67  1728 1.00
#> P_sim[50]                 1.00  2205 1.00
#> P_sim[51]                 0.50  1899 1.00
#> P_sim[52]                 0.50  1925 1.00
#> P_sim[53]                 0.67  1875 1.00
#> P_sim[54]                 0.50  1869 1.00
#> P_sim[55]                 1.00  1988 1.00
#> P_sim[56]                 0.75  1997 1.00
#> P_sim[57]                 0.67  2091 1.00
#> P_sim[58]                 1.00  1960 1.00
#> P_sim[59]                 0.45  1765 1.00
#> P_sim[60]                 0.60  2106 1.00
#> P_sim[61]                 1.00  1598 1.00
#> P_sim[62]                 1.00  1760 1.00
#> P_sim[63]                 0.60  1797 1.00
#> P_sim[64]                 0.50  2004 1.00
#> P_sim[65]                 1.00  2024 1.00
#> P_sim[66]                 1.00  2003 1.00
#> P_sim[67]                 1.00  1769 1.00
#> P_sim[68]                 0.60  1960 1.00
#> P_sim[69]                 0.67  1983 1.00
#> P_sim[70]                 1.00  1988 1.00
#> P_sim[71]                 1.00  1917 1.00
#> P_sim[72]                 1.00  1711 1.00
#> logLikelihood[1]         -0.10   777 1.00
#> logLikelihood[2]         -0.15   777 1.00
#> logLikelihood[3]         -1.15   773 1.00
#> logLikelihood[4]         -0.34   777 1.00
#> logLikelihood[5]         -0.53   777 1.00
#> logLikelihood[6]         -0.40   777 1.00
#> logLikelihood[7]         -0.70   777 1.00
#> logLikelihood[8]         -0.42   777 1.00
#> logLikelihood[9]         -0.55   777 1.00
#> logLikelihood[10]        -0.70   777 1.00
#> logLikelihood[11]        -1.02   769 1.00
#> logLikelihood[12]        -0.99   777 1.00
#> logLikelihood[13]        -0.57   777 1.00
#> logLikelihood[14]        -1.03   777 1.00
#> logLikelihood[15]        -0.57   777 1.00
#> logLikelihood[16]        -0.75   777 1.00
#> logLikelihood[17]        -0.60   777 1.00
#> logLikelihood[18]        -0.95   777 1.00
#> logLikelihood[19]        -0.50   777 1.00
#> logLikelihood[20]        -0.47   777 1.00
#> logLikelihood[21]        -0.25   777 1.00
#> logLikelihood[22]        -1.23   777 1.00
#> logLikelihood[23]        -1.01   777 1.00
#> logLikelihood[24]        -0.99   777 1.00
#> logLikelihood[25]        -1.47   777 1.00
#> logLikelihood[26]        -0.95   536 1.00
#> logLikelihood[27]        -0.93   731 1.00
#> logLikelihood[28]        -0.74   777 1.00
#> logLikelihood[29]        -0.77   777 1.00
#> logLikelihood[30]        -0.95   610 1.00
#> logLikelihood[31]        -0.95   653 1.00
#> logLikelihood[32]        -1.03   777 1.00
#> logLikelihood[33]        -0.95   737 1.00
#> logLikelihood[34]        -0.80   777 1.00
#> logLikelihood[35]        -0.95   725 1.00
#> logLikelihood[36]        -1.86   773 1.00
#> logLikelihood[37]        -0.94   701 1.00
#> logLikelihood[38]        -0.93   770 1.00
#> logLikelihood[39]        -0.46   777 1.00
#> logLikelihood[40]        -1.22   765 1.00
#> logLikelihood[41]        -1.96   775 1.00
#> logLikelihood[42]        -0.99   777 1.00
#> logLikelihood[43]        -0.97   772 1.00
#> logLikelihood[44]        -1.61   775 1.00
#> logLikelihood[45]        -0.40   777 1.00
#> logLikelihood[46]        -2.27   775 1.00
#> logLikelihood[47]        -3.06   775 1.00
#> logLikelihood[48]        -2.17   775 1.00
#> logLikelihood[49]        -0.44   777 1.00
#> logLikelihood[50]        -1.47   775 1.00
#> logLikelihood[51]        -2.17   774 1.00
#> logLikelihood[52]        -0.62   777 1.00
#> logLikelihood[53]        -0.88   770 1.00
#> logLikelihood[54]        -2.07   774 1.00
#> logLikelihood[55]        -0.16   777 1.00
#> logLikelihood[56]        -3.04   775 1.00
#> logLikelihood[57]        -1.92   775 1.00
#> logLikelihood[58]        -0.17   777 1.00
#> logLikelihood[59]        -1.19   775 1.00
#> logLikelihood[60]        -0.89   777 1.00
#> logLikelihood[61]        -1.31   775 1.00
#> logLikelihood[62]        -0.92   773 1.00
#> logLikelihood[63]        -2.19   774 1.00
#> logLikelihood[64]        -1.17   652 1.00
#> logLikelihood[65]        -1.25   775 1.00
#> logLikelihood[66]        -1.24   775 1.00
#> logLikelihood[67]        -0.40   777 1.00
#> logLikelihood[68]        -1.18   769 1.00
#> logLikelihood[69]        -1.67   774 1.00
#> logLikelihood[70]        -1.19   775 1.00
#> logLikelihood[71]        -1.18   775 1.00
#> logLikelihood[72]        -1.13   775 1.00
#> lp__                   -155.55   471 1.01
#> 
#> Samples were drawn using NUTS(diag_e) at Tue Apr 23 16:25:30 2024.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).