Sets uniform distribution parameters for sampling

Sets uniform distribution parameters for sampling

Usage

sf_uniform(min = 0, max = 10)

Arguments

min

Minimum value of the random variable of the uniform distribution

max

Maximum value of the random variable of the uniform distribution

Value

List with specified statistics and name of the model

Examples

my_prior <- sf_uniform()