Fit selected model to the specified seroprevalence survey data

This function fits the specified model foi_model to the serological survey data serodata by means of sampling. The function determines whether the corresponding stan model object needs to be compiled by rstan.

Usage

fit_seromodel(
  serodata,
  foi_model = c("constant", "tv_normal_log", "tv_normal"),
  iter = 1000,
  thin = 2,
  adapt_delta = 0.9,
  max_treedepth = 10,
  chains = 4,
  seed = 12345,
  ...
)

Arguments

serodata

A data frame containing the data from a seroprevalence survey. This data frame must contain at least the following columns:

total

Number of samples for each age group

counts

Number of positive samples for each age group

tsur

Year in which the survey took place

age_mean_f

Floor value of the average between age_min and age_max

sample_size

The size of the sample

birth_year

The year in which the individuals of each age group were born

The last six columns can be added to serodata by means of the function prepare_serodata().

foi_model

Name of the selected model. Current version provides three options:

"constant"

Runs a constant model

"tv_normal"

Runs a normal model

"tv_normal_log"

Runs a normal logarithmic model

iter

Number of interactions for each chain including the warmup. iter in sampling.

thin

Positive integer specifying the period for saving samples. thin in sampling.

adapt_delta

Real number between 0 and 1 that represents the target average acceptance probability. Increasing the value of adapt_delta will result in a smaller step size and fewer divergences. For further details refer to the control parameter in sampling or here.

max_treedepth

Maximum tree depth for the binary tree used in the NUTS stan sampler. For further details refer to the control parameter in sampling.

chains

Number of Markov chains for sampling. For further details refer to the chains parameter in sampling.

seed

For further details refer to the seed parameter in sampling.

...

Additional parameters for sampling.

Value

seromodel_object. stanfit object returned by the function sampling

Examples

data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_fit <- fit_seromodel(
  serodata = serodata,
  foi_model = "constant"
)