Function that generates a data.frame containing the age of each cohort corresponding to each birth year exluding the year of the survey.

This function generates a data.frame containing the age of each cohort corresponding to each birth_year excluding the year of the survey, for which the cohort age is still 0. specified serological survey data serodata excluding the year of the survey.

Usage

get_cohort_ages(serodata)

Arguments

serodata

A data frame containing the data from a seroprevalence survey. This data frame must contain the following columns:

survey	survey Label of the current survey

total	Number of samples for each age group

counts	Number of positive samples for each age group

age_min	age_min

age_max	age_max

tsur	Year in which the survey took place

country	The country where the survey took place

test	The type of test taken

antibody	antibody

age_mean_f	Floor value of the average between age_min and age_max

sample_size	The size of the sample

birth_year	The year in which the individuals of each age group were bornt

prev_obs	Observed prevalence

prev_obs_lower	Lower limit of the confidence interval for the observed prevalence

prev_obs_upper	Upper limit of the confidence interval for the observed prevalence

The last six colums can be added to serodata by means of the function prepare_serodata.

Value

cohort_ages. A data.frame containing the age of each cohort corresponding to each birth year

Examples

data(chagas2012)
serodata <- prepare_serodata(serodata = chagas2012, alpha = 0.05)
cohort_ages <- get_cohort_ages(serodata = serodata)