Generate data frame containing the confidence interval based on a force-of-infection fitting

This function computes the corresponding binomial confidence intervals for the obtained prevalence based on a fitting of the force-of-infection foi for plotting an analysis purposes.

Usage

get_prev_expanded(foi, serodata, alpha = 0.05, bin_data = FALSE, bin_step = 5)

Arguments

foi

Object containing the information of the force-of-infection. It is obtained from rstan::extract(seromodel_object$seromodel, "foi", inc_warmup = FALSE)[[1]].

serodata

A data frame containing the data from a seroprevalence survey. This data frame must contain at least the following columns:

total

Number of samples for each age group

counts

Number of positive samples for each age group

tsur

Year in which the survey took place

age_mean_f

Floor value of the average between age_min and age_max

sample_size

The size of the sample

birth_year

The year in which the individuals of each age group were born

The last six columns can be added to serodata by means of the function prepare_serodata().

alpha

Probability threshold for statistical significance used for both the binomial confidence interval, and the lower and upper quantiles of the estimated prevalence.

bin_data

If TRUE, serodata is binned by means of prepare_bin_data. Otherwise, age groups are kept as originally input.

bin_step

Integer specifying the age groups bin size to be used when bin_data is set to TRUE.

Value

prev_final. The expanded prevalence data. This is used for plotting purposes in the visualization module.

Examples

data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- run_seromodel(
  serodata = serodata,
  foi_model = "constant"
)
#> serofoi model constant finished running ------
#>           [,1]              
#> foi_model "constant"        
#> dataset   "COL-035-93"      
#> country   "COL"             
#> year      "2012"            
#> test      "ELISA"           
#> antibody  "IgG anti-T.cruzi"
#> n_sample  "747"             
#> n_agec    "72"              
#> n_iter    "1000"            
#> elpd      "-92.73"          
#> se        "6.42"            
#> converged "Yes"             
foi <- rstan::extract(seromodel_object, "foi")[[1]]
get_prev_expanded(foi, serodata)
#>    age predicted_prev predicted_prev_lower predicted_prev_upper   prev_obs
#> 1    1    0.003863759          0.003065157          0.004880402 0.00000000
#> 2    2    0.007712589          0.006120920          0.009736986 0.00000000
#> 3    3    0.011546548          0.009167316          0.014569868 0.02857143
#> 4    4    0.015365693          0.012204374          0.019379164 0.00000000
#> 5    5    0.019170082          0.015232123          0.024164988 0.00000000
#> 6    6    0.022959772          0.018250592          0.028927456 0.00000000
#> 7    7    0.026734820          0.021259808          0.033666680 0.00000000
#> 8    8    0.030495282          0.024259801          0.038382776 0.00000000
#> 9    9    0.034241214          0.027250598          0.043075855 0.00000000
#> 10  10    0.037972673          0.030232228          0.047746030 0.00000000
#> 11  11    0.041689714          0.033204719          0.052393412 0.07692308
#> 12  12    0.045392394          0.036168099          0.057018114 0.00000000
#> 13  13    0.049080767          0.039122395          0.061620245 0.00000000
#> 14  14    0.052754889          0.042067637          0.066199915 0.00000000
#> 15  15    0.056414816          0.045003850          0.070757235 0.00000000
#> 16  16    0.060060601          0.047931064          0.075292314 0.00000000
#> 17  17    0.063692300          0.050849305          0.079805260 0.00000000
#> 18  18    0.067309967          0.053758601          0.084296180 0.00000000
#> 19  19    0.070913656          0.056658980          0.088765183 0.00000000
#> 20  20    0.074503422          0.059550469          0.093212376 0.00000000
#> 21  21    0.078079317          0.062433095          0.097637864 0.00000000
#> 22  22    0.081641396          0.065306885          0.102041755 0.00000000
#> 23  23    0.085189712          0.068171866          0.106424152 0.00000000
#> 24  24    0.088724318          0.071028066          0.110785162 0.00000000
#> 25  25    0.092245267          0.073875512          0.115124888 0.00000000
#> 26  26    0.095752612          0.076714229          0.119443435 0.09090909
#> 27  27    0.099246406          0.079544245          0.123740905 0.12500000
#> 28  28    0.102726701          0.082365587          0.128017402 0.00000000
#> 29  29    0.106193548          0.085178281          0.132273028 0.00000000
#> 30  30    0.109647000          0.087982353          0.136507885 0.10000000
#> 31  31    0.113087109          0.090777831          0.140722074 0.10000000
#> 32  32    0.116513927          0.093564740          0.144915696 0.00000000
#> 33  33    0.119927504          0.096343107          0.149088851 0.09090909
#> 34  34    0.123327891          0.099112958          0.153241640 0.00000000
#> 35  35    0.126715141          0.101874318          0.157374161 0.10000000
#> 36  36    0.130089303          0.104627215          0.161486514 0.30000000
#> 37  37    0.133450428          0.107371673          0.165578798 0.11111111
#> 38  38    0.136798566          0.110107720          0.169651109 0.25000000
#> 39  39    0.140133768          0.112835380          0.173703545 0.00000000
#> 40  40    0.143456083          0.115554679          0.177736205 0.22222222
#> 41  41    0.146765562          0.118265643          0.181749183 0.50000000
#> 42  42    0.150062254          0.120968298          0.185742576 0.00000000
#> 43  43    0.153346209          0.123662668          0.189716480 0.33333333
#> 44  44    0.156617474          0.126348780          0.193670989 1.00000000
#> 45  45    0.159876101          0.129026659          0.197606199 0.00000000
#> 46  46    0.163122137          0.131696329          0.201522204 0.66666667
#> 47  47    0.166355631          0.134357816          0.205419097         NA
#> 48  48    0.169576631          0.137011146          0.209296971 1.00000000
#> 49  49    0.172785187          0.139656343          0.213155920 0.66666667
#> 50  50    0.175981345          0.142293431          0.216996036 0.00000000
#> 51  51    0.179165154          0.144922437          0.220817410 1.00000000
#> 52  52    0.182336662          0.147543384          0.224620135 0.50000000
#> 53  53    0.185495916          0.150156298          0.228404301 0.00000000
#> 54  54    0.188642963          0.152761203          0.232169998 0.33333333
#> 55  55    0.191777851          0.155358123          0.235917318 0.50000000
#> 56  56    0.194900626          0.157947083          0.239646348 0.00000000
#> 57  57    0.198011336          0.160528108          0.243357180 0.75000000
#> 58  58    0.201110026          0.163101222          0.247049902 0.66666667
#> 59  59    0.204196744          0.165666448          0.250724601 0.00000000
#> 60  60    0.207271536          0.168223812          0.254381366 0.09090909
#> 61  61    0.210334447          0.170773337          0.258020285 0.00000000
#> 62  62    0.213385524          0.173315047          0.261641445 1.00000000
#> 63  63    0.216424813          0.175848967          0.265244932 0.50000000
#> 64  64    0.219452358          0.178375119          0.268830832 0.60000000
#> 65  65    0.222468206          0.180893529          0.272399232 0.25000000
#> 66  66    0.225472401          0.183404219          0.275950216 1.00000000
#> 67  67    0.228464989          0.185907214          0.279483871 1.00000000
#> 68  68    0.231446014          0.188402536          0.283000279 0.00000000
#> 69  69    0.234415521          0.190890210          0.286499526 0.40000000
#> 70  70    0.237373554          0.193370259          0.289981696 0.66666667
#> 71  71    0.240320159          0.195842706          0.293446871 1.00000000
#> 72  72    0.243255378          0.198307575          0.296895134 1.00000000
#> 73  73    0.246179257          0.200764888          0.300326569         NA
#> 74  74    0.249091838          0.203214670          0.303741257         NA
#> 75  75    0.251993166          0.205656942          0.307139280         NA
#> 76  76    0.254883284          0.208091729          0.310520719         NA
#> 77  77    0.257762235          0.210519052          0.313885655 1.00000000
#>    prev_obs_lower prev_obs_upper total counts     survey
#> 1    0.0000000000     0.10281792    34      0 COL-035-93
#> 2    0.0000000000     0.13718517    25      0 COL-035-93
#> 3    0.0007231044     0.14917208    35      1 COL-035-93
#> 4    0.0000000000     0.11944487    29      0 COL-035-93
#> 5    0.0000000000     0.09739376    36      0 COL-035-93
#> 6    0.0000000000     0.14818513    23      0 COL-035-93
#> 7    0.0000000000     0.10281792    34      0 COL-035-93
#> 8    0.0000000000     0.18530197    18      0 COL-035-93
#> 9    0.0000000000     0.16109762    21      0 COL-035-93
#> 10   0.0000000000     0.14247360    24      0 COL-035-93
#> 11   0.0019456285     0.36029744    13      1 COL-035-93
#> 12   0.0000000000     0.12343612    28      0 COL-035-93
#> 13   0.0000000000     0.21801936    15      0 COL-035-93
#> 14   0.0000000000     0.13718517    25      0 COL-035-93
#> 15   0.0000000000     0.24705264    13      0 COL-035-93
#> 16   0.0000000000     0.20590721    16      0 COL-035-93
#> 17   0.0000000000     0.26464847    12      0 COL-035-93
#> 18   0.0000000000     0.18530197    18      0 COL-035-93
#> 19   0.0000000000     0.33626712     9      0 COL-035-93
#> 20   0.0000000000     0.36941665     8      0 COL-035-93
#> 21   0.0000000000     0.60236464     4      0 COL-035-93
#> 22   0.0000000000     0.17646691    19      0 COL-035-93
#> 23   0.0000000000     0.21801936    15      0 COL-035-93
#> 24   0.0000000000     0.23163576    14      0 COL-035-93
#> 25   0.0000000000     0.16843347    20      0 COL-035-93
#> 26   0.0022989722     0.41277992    11      1 COL-035-93
#> 27   0.0031597235     0.52650967     8      1 COL-035-93
#> 28   0.0000000000     0.33626712     9      0 COL-035-93
#> 29   0.0000000000     0.33626712     9      0 COL-035-93
#> 30   0.0025285785     0.44501612    10      1 COL-035-93
#> 31   0.0025285785     0.44501612    10      1 COL-035-93
#> 32   0.0000000000     0.28491415    11      0 COL-035-93
#> 33   0.0022989722     0.41277992    11      1 COL-035-93
#> 34   0.0000000000     0.36941665     8      0 COL-035-93
#> 35   0.0025285785     0.44501612    10      1 COL-035-93
#> 36   0.0667395112     0.65245285    10      3 COL-035-93
#> 37   0.0028091367     0.48249651     9      1 COL-035-93
#> 38   0.0063094632     0.80587955     4      1 COL-035-93
#> 39   0.0000000000     0.60236464     4      0 COL-035-93
#> 40   0.0281449735     0.60009357     9      2 COL-035-93
#> 41   0.0675859865     0.93241401     4      2 COL-035-93
#> 42   0.0000000000     0.36941665     8      0 COL-035-93
#> 43   0.0084037587     0.90570068     3      1 COL-035-93
#> 44   0.0250000000     1.00000000     1      1 COL-035-93
#> 45   0.0000000000     0.70759823     3      0 COL-035-93
#> 46   0.0942993241     0.99159624     3      2 COL-035-93
#> 47             NA             NA    NA     NA COL-035-93
#> 48   0.1581138830     1.00000000     2      2 COL-035-93
#> 49   0.0942993241     0.99159624     3      2 COL-035-93
#> 50   0.0000000000     0.70759823     3      0 COL-035-93
#> 51   0.0250000000     1.00000000     1      1 COL-035-93
#> 52   0.1181172488     0.88188275     6      3 COL-035-93
#> 53   0.0000000000     0.60236464     4      0 COL-035-93
#> 54   0.0084037587     0.90570068     3      1 COL-035-93
#> 55   0.1181172488     0.88188275     6      3 COL-035-93
#> 56   0.0000000000     0.97500000     1      0 COL-035-93
#> 57   0.1941204497     0.99369054     4      3 COL-035-93
#> 58   0.0942993241     0.99159624     3      2 COL-035-93
#> 59   0.0000000000     0.97500000     1      0 COL-035-93
#> 60   0.0022989722     0.41277992    11      1 COL-035-93
#> 61   0.0000000000     0.52182375     5      0 COL-035-93
#> 62   0.0250000000     1.00000000     1      1 COL-035-93
#> 63   0.0125791171     0.98742088     2      1 COL-035-93
#> 64   0.1466327996     0.94725505     5      3 COL-035-93
#> 65   0.0318540262     0.65085579     8      2 COL-035-93
#> 66   0.0250000000     1.00000000     1      1 COL-035-93
#> 67   0.0250000000     1.00000000     1      1 COL-035-93
#> 68   0.0000000000     0.84188612     2      0 COL-035-93
#> 69   0.0527449505     0.85336720     5      2 COL-035-93
#> 70   0.0942993241     0.99159624     3      2 COL-035-93
#> 71   0.0250000000     1.00000000     1      1 COL-035-93
#> 72   0.0250000000     1.00000000     1      1 COL-035-93
#> 73             NA             NA    NA     NA COL-035-93
#> 74             NA             NA    NA     NA COL-035-93
#> 75             NA             NA    NA     NA COL-035-93
#> 76             NA             NA    NA     NA COL-035-93
#> 77   0.0250000000     1.00000000     1      1 COL-035-93