Function that generates an object containing the confidence interval based on a Force-of-Infection fitting

This function computes the corresponding binomial confidence intervals for the obtained prevalence based on a fitting of the Force-of-Infection foi for plotting an analysis purposes.

Usage

get_prev_expanded(
  foi,
  serodata,
  predicted_prev_lower_quantile = 0.1,
  predicted_prev_upper_quantile = 0.9,
  bin_data = FALSE
)

Arguments

foi

Object containing the information of the force of infection. It is obtained from rstan::extract(seromodel_object$seromodel, "foi", inc_warmup = FALSE)[[1]].

serodata

A data frame containing the data from a seroprevalence survey. This data frame must contain the following columns:

survey

survey Label of the current survey

total

Number of samples for each age group

counts

Number of positive samples for each age group

age_min

age_min

age_max

age_max

tsur

Year in which the survey took place

country

The country where the survey took place

test

The type of test taken

antibody

antibody

age_mean_f

Floor value of the average between age_min and age_max

sample_size

The size of the sample

birth_year

The year in which the individuals of each age group were born

prev_obs

Observed prevalence

prev_obs_lower

Lower limit of the confidence interval for the observed prevalence

prev_obs_upper

Upper limit of the confidence interval for the observed prevalence

The last six columns can be added to serodata by means of the function prepare_serodata().

predicted_prev_lower_quantile

Float specifying the lower quantile to be used in predicted_prev calculation

predicted_prev_upper_quantile

Float specifying the upper quantile to be used in predicted_prev calculation

bin_data

If TRUE, serodata is binned by means of prepare_bin_data. Otherwise, age groups are kept as originally input.

Value

prev_final. The expanded prevalence data. This is used for plotting purposes in the visualization module.

Examples

data(chagas2012)
serodata <- prepare_serodata(chagas2012)
seromodel_object <- run_seromodel(
  serodata = serodata,
  foi_model = "constant"
)
#> serofoi model constant finished running ------
#>           [,1]              
#> foi_model "constant"        
#> dataset   "COL-035-93"      
#> country   "COL"             
#> year      "2012"            
#> test      "ELISA"           
#> antibody  "IgG anti-T.cruzi"
#> n_sample  "747"             
#> n_agec    "72"              
#> n_iter    "1000"            
#> elpd      "-92.75"          
#> se        "6.4"             
#> converged "Yes"             
foi <- rstan::extract(seromodel_object, "foi")[[1]]
get_prev_expanded(foi, serodata)
#>    age predicted_prev predicted_prev_lower predicted_prev_upper   prev_obs
#> 1    1    0.003893214          0.003242693          0.004664603 0.00000000
#> 2    2    0.007771271          0.006474871          0.009307447 0.00000000
#> 3    3    0.011634230          0.009696568          0.013928634 0.02857143
#> 4    4    0.015482149          0.012907818          0.018528265 0.00000000
#> 5    5    0.019315088          0.016108654          0.023106440 0.00000000
#> 6    6    0.023133104          0.019299112          0.027663260 0.00000000
#> 7    7    0.026936256          0.022479224          0.032198825 0.00000000
#> 8    8    0.030724601          0.025649023          0.036713233 0.00000000
#> 9    9    0.034498198          0.028808544          0.041206583 0.00000000
#> 10  10    0.038257103          0.031957820          0.045678973 0.00000000
#> 11  11    0.042001374          0.035096884          0.050130501 0.07692308
#> 12  12    0.045731067          0.038225768          0.054561265 0.00000000
#> 13  13    0.049446241          0.041344507          0.058971361 0.00000000
#> 14  14    0.053146950          0.044453132          0.063360885 0.00000000
#> 15  15    0.056833251          0.047551677          0.067729934 0.00000000
#> 16  16    0.060505201          0.050640174          0.072078604 0.00000000
#> 17  17    0.064162855          0.053718657          0.076406988 0.00000000
#> 18  18    0.067806270          0.056787157          0.080715183 0.00000000
#> 19  19    0.071435499          0.059845706          0.085003281 0.00000000
#> 20  20    0.075050600          0.062894338          0.089271377 0.00000000
#> 21  21    0.078651626          0.065933084          0.093519564 0.00000000
#> 22  22    0.082238632          0.068961976          0.097747935 0.00000000
#> 23  23    0.085811673          0.071981046          0.101956582 0.00000000
#> 24  24    0.089370804          0.074990327          0.106145597 0.00000000
#> 25  25    0.092916078          0.077989849          0.110315073 0.00000000
#> 26  26    0.096447550          0.080979645          0.114465099 0.09090909
#> 27  27    0.099965273          0.083959746          0.118595768 0.12500000
#> 28  28    0.103469301          0.086930183          0.122707168 0.00000000
#> 29  29    0.106959687          0.089890988          0.126799390 0.00000000
#> 30  30    0.110436484          0.092842192          0.130872524 0.10000000
#> 31  31    0.113899745          0.095783826          0.134926658 0.10000000
#> 32  32    0.117349523          0.098715921          0.138961881 0.00000000
#> 33  33    0.120785870          0.101638509          0.142978282 0.09090909
#> 34  34    0.124208839          0.104551619          0.146975948 0.00000000
#> 35  35    0.127618481          0.107455283          0.150954966 0.10000000
#> 36  36    0.131014849          0.110349532          0.154915423 0.30000000
#> 37  37    0.134397994          0.113234395          0.158857407 0.11111111
#> 38  38    0.137767968          0.116109903          0.162781002 0.25000000
#> 39  39    0.141124822          0.118976088          0.166686296 0.00000000
#> 40  40    0.144468607          0.121832978          0.170573373 0.22222222
#> 41  41    0.147799374          0.124680604          0.174442319 0.50000000
#> 42  42    0.151117173          0.127518996          0.178293217 0.00000000
#> 43  43    0.154422055          0.130348183          0.182126152 0.33333333
#> 44  44    0.157714071          0.133168197          0.185941209 1.00000000
#> 45  45    0.160993271          0.135979067          0.189738469 0.00000000
#> 46  46    0.164259703          0.138780821          0.193518017 0.66666667
#> 47  47    0.167513419          0.141573490          0.197279935         NA
#> 48  48    0.170754468          0.144357104          0.201024305 1.00000000
#> 49  49    0.173982898          0.147131691          0.204751209 0.66666667
#> 50  50    0.177198759          0.149897281          0.208460728 0.00000000
#> 51  51    0.180402101          0.152653903          0.212152944 1.00000000
#> 52  52    0.183592971          0.155401586          0.215827937 0.50000000
#> 53  53    0.186771418          0.158140359          0.219485788 0.00000000
#> 54  54    0.189937491          0.160870252          0.223126577 0.33333333
#> 55  55    0.193091237          0.163591292          0.226750382 0.50000000
#> 56  56    0.196232706          0.166303508          0.230357284 0.00000000
#> 57  57    0.199361944          0.169006930          0.233947361 0.75000000
#> 58  58    0.202478999          0.171701585          0.237520692 0.66666667
#> 59  59    0.205583919          0.174387503          0.241077355 0.00000000
#> 60  60    0.208676751          0.177064710          0.244617427 0.09090909
#> 61  61    0.211757542          0.179733237          0.248140987 0.00000000
#> 62  62    0.214826338          0.182393110          0.251648110 1.00000000
#> 63  63    0.217883187          0.185044358          0.255138874 0.50000000
#> 64  64    0.220928135          0.187687009          0.258613355 0.60000000
#> 65  65    0.223961229          0.190321090          0.262071629 0.25000000
#> 66  66    0.226982514          0.192946630          0.265513771 1.00000000
#> 67  67    0.229992036          0.195563657          0.268939857 1.00000000
#> 68  68    0.232989842          0.198172197          0.272349962 0.00000000
#> 69  69    0.235975977          0.200772278          0.275744160 0.40000000
#> 70  70    0.238950486          0.203363928          0.279122525 0.66666667
#> 71  71    0.241913414          0.205947174          0.282485132 1.00000000
#> 72  72    0.244864808          0.208522043          0.285832054 1.00000000
#> 73  73    0.247804710          0.211088563          0.289163363         NA
#> 74  74    0.250733168          0.213646761          0.292479133         NA
#> 75  75    0.253650224          0.216196663          0.295779437         NA
#> 76  76    0.256555923          0.218738296          0.299064345         NA
#> 77  77    0.259450310          0.221271688          0.302333931 1.00000000
#> 78  78    0.262333428          0.223796865          0.305588266         NA
#> 79  79    0.265205322          0.226313853          0.308827421         NA
#> 80  80    0.268066035          0.228822680          0.312051466         NA
#>    prev_obs_lower prev_obs_upper sample_by_age positives     survey
#> 1    0.0000000000     0.10281792            34         0 COL-035-93
#> 2    0.0000000000     0.13718517            25         0 COL-035-93
#> 3    0.0007231044     0.14917208            35         1 COL-035-93
#> 4    0.0000000000     0.11944487            29         0 COL-035-93
#> 5    0.0000000000     0.09739376            36         0 COL-035-93
#> 6    0.0000000000     0.14818513            23         0 COL-035-93
#> 7    0.0000000000     0.10281792            34         0 COL-035-93
#> 8    0.0000000000     0.18530197            18         0 COL-035-93
#> 9    0.0000000000     0.16109762            21         0 COL-035-93
#> 10   0.0000000000     0.14247360            24         0 COL-035-93
#> 11   0.0019456285     0.36029744            13         1 COL-035-93
#> 12   0.0000000000     0.12343612            28         0 COL-035-93
#> 13   0.0000000000     0.21801936            15         0 COL-035-93
#> 14   0.0000000000     0.13718517            25         0 COL-035-93
#> 15   0.0000000000     0.24705264            13         0 COL-035-93
#> 16   0.0000000000     0.20590721            16         0 COL-035-93
#> 17   0.0000000000     0.26464847            12         0 COL-035-93
#> 18   0.0000000000     0.18530197            18         0 COL-035-93
#> 19   0.0000000000     0.33626712             9         0 COL-035-93
#> 20   0.0000000000     0.36941665             8         0 COL-035-93
#> 21   0.0000000000     0.60236464             4         0 COL-035-93
#> 22   0.0000000000     0.17646691            19         0 COL-035-93
#> 23   0.0000000000     0.21801936            15         0 COL-035-93
#> 24   0.0000000000     0.23163576            14         0 COL-035-93
#> 25   0.0000000000     0.16843347            20         0 COL-035-93
#> 26   0.0022989722     0.41277992            11         1 COL-035-93
#> 27   0.0031597235     0.52650967             8         1 COL-035-93
#> 28   0.0000000000     0.33626712             9         0 COL-035-93
#> 29   0.0000000000     0.33626712             9         0 COL-035-93
#> 30   0.0025285785     0.44501612            10         1 COL-035-93
#> 31   0.0025285785     0.44501612            10         1 COL-035-93
#> 32   0.0000000000     0.28491415            11         0 COL-035-93
#> 33   0.0022989722     0.41277992            11         1 COL-035-93
#> 34   0.0000000000     0.36941665             8         0 COL-035-93
#> 35   0.0025285785     0.44501612            10         1 COL-035-93
#> 36   0.0667395112     0.65245285            10         3 COL-035-93
#> 37   0.0028091367     0.48249651             9         1 COL-035-93
#> 38   0.0063094632     0.80587955             4         1 COL-035-93
#> 39   0.0000000000     0.60236464             4         0 COL-035-93
#> 40   0.0281449735     0.60009357             9         2 COL-035-93
#> 41   0.0675859865     0.93241401             4         2 COL-035-93
#> 42   0.0000000000     0.36941665             8         0 COL-035-93
#> 43   0.0084037587     0.90570068             3         1 COL-035-93
#> 44   0.0250000000     1.00000000             1         1 COL-035-93
#> 45   0.0000000000     0.70759823             3         0 COL-035-93
#> 46   0.0942993241     0.99159624             3         2 COL-035-93
#> 47             NA             NA            NA        NA COL-035-93
#> 48   0.1581138830     1.00000000             2         2 COL-035-93
#> 49   0.0942993241     0.99159624             3         2 COL-035-93
#> 50   0.0000000000     0.70759823             3         0 COL-035-93
#> 51   0.0250000000     1.00000000             1         1 COL-035-93
#> 52   0.1181172488     0.88188275             6         3 COL-035-93
#> 53   0.0000000000     0.60236464             4         0 COL-035-93
#> 54   0.0084037587     0.90570068             3         1 COL-035-93
#> 55   0.1181172488     0.88188275             6         3 COL-035-93
#> 56   0.0000000000     0.97500000             1         0 COL-035-93
#> 57   0.1941204497     0.99369054             4         3 COL-035-93
#> 58   0.0942993241     0.99159624             3         2 COL-035-93
#> 59   0.0000000000     0.97500000             1         0 COL-035-93
#> 60   0.0022989722     0.41277992            11         1 COL-035-93
#> 61   0.0000000000     0.52182375             5         0 COL-035-93
#> 62   0.0250000000     1.00000000             1         1 COL-035-93
#> 63   0.0125791171     0.98742088             2         1 COL-035-93
#> 64   0.1466327996     0.94725505             5         3 COL-035-93
#> 65   0.0318540262     0.65085579             8         2 COL-035-93
#> 66   0.0250000000     1.00000000             1         1 COL-035-93
#> 67   0.0250000000     1.00000000             1         1 COL-035-93
#> 68   0.0000000000     0.84188612             2         0 COL-035-93
#> 69   0.0527449505     0.85336720             5         2 COL-035-93
#> 70   0.0942993241     0.99159624             3         2 COL-035-93
#> 71   0.0250000000     1.00000000             1         1 COL-035-93
#> 72   0.0250000000     1.00000000             1         1 COL-035-93
#> 73             NA             NA            NA        NA COL-035-93
#> 74             NA             NA            NA        NA COL-035-93
#> 75             NA             NA            NA        NA COL-035-93
#> 76             NA             NA            NA        NA COL-035-93
#> 77   0.0250000000     1.00000000             1         1 COL-035-93
#> 78             NA             NA            NA        NA COL-035-93
#> 79             NA             NA            NA        NA COL-035-93
#> 80             NA             NA            NA        NA COL-035-93