Build dataframe containing the R-hat estimates for a given serological model

This function relies on rhat to extract the R-hat estimates of the serological model object seromodel_object and returns a table a dataframe with the estimates for each year of birth.

Usage

get_table_rhats(seromodel_object, cohort_ages)

Arguments

seromodel_object

Stanfit object containing the results of fitting a model by means of run_seromodel.

cohort_ages

A data frame containing the age of each cohort corresponding to each birth year.

Value

rhats table

Examples

data(chagas2012)
serodata <- prepare_serodata(serodata = chagas2012)
model_constant <- fit_seromodel(
  serodata = serodata,
  foi_model = "constant",
  iter = 1500
)
cohort_ages <- get_cohort_ages(serodata)
get_table_rhats(
  seromodel_object = model_constant,
  cohort_ages = cohort_ages
)
#>         year     rhat
#> foi[1]  1935 1.001341
#> foi[2]  1936 1.001341
#> foi[3]  1937 1.001341
#> foi[4]  1938 1.001341
#> foi[5]  1939 1.001341
#> foi[6]  1940 1.001341
#> foi[7]  1941 1.001341
#> foi[8]  1942 1.001341
#> foi[9]  1943 1.001341
#> foi[10] 1944 1.001341
#> foi[11] 1945 1.001341
#> foi[12] 1946 1.001341
#> foi[13] 1947 1.001341
#> foi[14] 1948 1.001341
#> foi[15] 1949 1.001341
#> foi[16] 1950 1.001341
#> foi[17] 1951 1.001341
#> foi[18] 1952 1.001341
#> foi[19] 1953 1.001341
#> foi[20] 1954 1.001341
#> foi[21] 1955 1.001341
#> foi[22] 1956 1.001341
#> foi[23] 1957 1.001341
#> foi[24] 1958 1.001341
#> foi[25] 1959 1.001341
#> foi[26] 1960 1.001341
#> foi[27] 1961 1.001341
#> foi[28] 1962 1.001341
#> foi[29] 1963 1.001341
#> foi[30] 1964 1.001341
#> foi[31] 1965 1.001341
#> foi[32] 1966 1.001341
#> foi[33] 1967 1.001341
#> foi[34] 1968 1.001341
#> foi[35] 1969 1.001341
#> foi[36] 1970 1.001341
#> foi[37] 1971 1.001341
#> foi[38] 1972 1.001341
#> foi[39] 1973 1.001341
#> foi[40] 1974 1.001341
#> foi[41] 1975 1.001341
#> foi[42] 1976 1.001341
#> foi[43] 1977 1.001341
#> foi[44] 1978 1.001341
#> foi[45] 1979 1.001341
#> foi[46] 1980 1.001341
#> foi[47] 1981 1.001341
#> foi[48] 1982 1.001341
#> foi[49] 1983 1.001341
#> foi[50] 1984 1.001341
#> foi[51] 1985 1.001341
#> foi[52] 1986 1.001341
#> foi[53] 1987 1.001341
#> foi[54] 1988 1.001341
#> foi[55] 1989 1.001341
#> foi[56] 1990 1.001341
#> foi[57] 1991 1.001341
#> foi[58] 1992 1.001341
#> foi[59] 1993 1.001341
#> foi[60] 1994 1.001341
#> foi[61] 1995 1.001341
#> foi[62] 1996 1.001341
#> foi[63] 1997 1.001341
#> foi[64] 1998 1.001341
#> foi[65] 1999 1.001341
#> foi[66] 2000 1.001341
#> foi[67] 2001 1.001341
#> foi[68] 2002 1.001341
#> foi[69] 2003 1.001341
#> foi[70] 2004 1.001341
#> foi[71] 2005 1.001341
#> foi[72] 2006 1.001341
#> foi[73] 2007 1.001341
#> foi[74] 2008 1.001341
#> foi[75] 2009 1.001341
#> foi[76] 2010 1.001341
#> foi[77] 2011 1.001341