Run specified stan model for the force-of-infection and estimate the seroprevalence based on the result of the fit

This function runs the specified model for the force-of-infection foi_model using the data from a seroprevalence survey serodata as the input data. See fit_seromodel for further details.

Usage

run_seromodel(
  serodata,
  foi_model = c("constant", "tv_normal_log", "tv_normal"),
  iter = 1000,
  thin = 2,
  adapt_delta = 0.9,
  max_treedepth = 10,
  chains = 4,
  seed = 12345,
  print_summary = TRUE,
  ...
)

Arguments

serodata

A data frame containing the data from a seroprevalence survey. This data frame must contain at least the following columns:

total

Number of samples for each age group

counts

Number of positive samples for each age group

tsur

Year in which the survey took place

age_mean_f

Floor value of the average between age_min and age_max

sample_size

The size of the sample

birth_year

The year in which the individuals of each age group were born

The last six columns can be added to serodata by means of the function prepare_serodata().

foi_model

Name of the selected model. Current version provides three options:

"constant"

Runs a constant model

"tv_normal"

Runs a normal model

"tv_normal_log"

Runs a normal logarithmic model

iter

Number of interactions for each chain including the warmup. iter in sampling.

thin

Positive integer specifying the period for saving samples. thin in sampling.

adapt_delta

Real number between 0 and 1 that represents the target average acceptance probability. Increasing the value of adapt_delta will result in a smaller step size and fewer divergences. For further details refer to the control parameter in sampling or here.

max_treedepth

Maximum tree depth for the binary tree used in the NUTS stan sampler. For further details refer to the control parameter in sampling.

chains

Number of Markov chains for sampling. For further details refer to the chains parameter in sampling.

seed

For further details refer to the seed parameter in sampling.

print_summary

Boolean. If TRUE, a table summarizing modelling results is printed.

...

Additional parameters for sampling.

Value

seromodel_object. An object containing relevant information about the implementation of the model. For further details refer to fit_seromodel.

Examples

data(chagas2012)
serodata <- prepare_serodata(chagas2012)
run_seromodel(
  serodata,
  foi_model = "constant"
)
#> serofoi model constant finished running ------
#>           [,1]              
#> foi_model "constant"        
#> dataset   "COL-035-93"      
#> country   "COL"             
#> year      "2012"            
#> test      "ELISA"           
#> antibody  "IgG anti-T.cruzi"
#> n_sample  "747"             
#> n_agec    "72"              
#> n_iter    "1000"            
#> elpd      "-92.75"          
#> se        "6.4"             
#> converged "Yes"             
#> Inference for Stan model: constant.
#> 4 chains, each with iter=1000; warmup=500; thin=2; 
#> post-warmup draws per chain=250, total post-warmup draws=1000.
#> 
#>                           mean se_mean   sd    2.5%     25%     50%     75%
#> lambda0                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> prob_infected[1]          0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> prob_infected[2]          0.01    0.00 0.00    0.01    0.01    0.01    0.01
#> prob_infected[3]          0.01    0.00 0.00    0.01    0.01    0.01    0.01
#> prob_infected[4]          0.02    0.00 0.00    0.01    0.01    0.02    0.02
#> prob_infected[5]          0.02    0.00 0.00    0.01    0.02    0.02    0.02
#> prob_infected[6]          0.02    0.00 0.00    0.02    0.02    0.02    0.03
#> prob_infected[7]          0.03    0.00 0.00    0.02    0.02    0.03    0.03
#> prob_infected[8]          0.03    0.00 0.00    0.02    0.03    0.03    0.03
#> prob_infected[9]          0.03    0.00 0.00    0.03    0.03    0.03    0.04
#> prob_infected[10]         0.04    0.00 0.01    0.03    0.03    0.04    0.04
#> prob_infected[11]         0.04    0.00 0.01    0.03    0.04    0.04    0.05
#> prob_infected[12]         0.05    0.00 0.01    0.04    0.04    0.05    0.05
#> prob_infected[13]         0.05    0.00 0.01    0.04    0.04    0.05    0.05
#> prob_infected[14]         0.05    0.00 0.01    0.04    0.05    0.05    0.06
#> prob_infected[15]         0.06    0.00 0.01    0.04    0.05    0.06    0.06
#> prob_infected[16]         0.06    0.00 0.01    0.05    0.05    0.06    0.07
#> prob_infected[17]         0.06    0.00 0.01    0.05    0.06    0.06    0.07
#> prob_infected[18]         0.07    0.00 0.01    0.05    0.06    0.07    0.07
#> prob_infected[19]         0.07    0.00 0.01    0.05    0.06    0.07    0.08
#> prob_infected[20]         0.08    0.00 0.01    0.06    0.07    0.08    0.08
#> prob_infected[21]         0.08    0.00 0.01    0.06    0.07    0.08    0.09
#> prob_infected[22]         0.08    0.00 0.01    0.06    0.07    0.08    0.09
#> prob_infected[23]         0.09    0.00 0.01    0.07    0.08    0.09    0.09
#> prob_infected[24]         0.09    0.00 0.01    0.07    0.08    0.09    0.10
#> prob_infected[25]         0.09    0.00 0.01    0.07    0.08    0.09    0.10
#> prob_infected[26]         0.10    0.00 0.01    0.07    0.09    0.10    0.10
#> prob_infected[27]         0.10    0.00 0.01    0.08    0.09    0.10    0.11
#> prob_infected[28]         0.10    0.00 0.01    0.08    0.09    0.10    0.11
#> prob_infected[29]         0.11    0.00 0.01    0.08    0.10    0.11    0.12
#> prob_infected[30]         0.11    0.00 0.01    0.09    0.10    0.11    0.12
#> prob_infected[31]         0.11    0.00 0.02    0.09    0.10    0.11    0.12
#> prob_infected[32]         0.12    0.00 0.02    0.09    0.11    0.12    0.13
#> prob_infected[33]         0.12    0.00 0.02    0.09    0.11    0.12    0.13
#> prob_infected[34]         0.12    0.00 0.02    0.10    0.11    0.12    0.13
#> prob_infected[35]         0.13    0.00 0.02    0.10    0.12    0.13    0.14
#> prob_infected[36]         0.13    0.00 0.02    0.10    0.12    0.13    0.14
#> prob_infected[37]         0.13    0.00 0.02    0.10    0.12    0.13    0.15
#> prob_infected[38]         0.14    0.00 0.02    0.11    0.13    0.14    0.15
#> prob_infected[39]         0.14    0.00 0.02    0.11    0.13    0.14    0.15
#> prob_infected[40]         0.15    0.00 0.02    0.11    0.13    0.14    0.16
#> prob_infected[41]         0.15    0.00 0.02    0.11    0.13    0.15    0.16
#> prob_infected[42]         0.15    0.00 0.02    0.12    0.14    0.15    0.16
#> prob_infected[43]         0.15    0.00 0.02    0.12    0.14    0.15    0.17
#> prob_infected[44]         0.16    0.00 0.02    0.12    0.14    0.16    0.17
#> prob_infected[45]         0.16    0.00 0.02    0.13    0.15    0.16    0.17
#> prob_infected[46]         0.16    0.00 0.02    0.13    0.15    0.16    0.18
#> prob_infected[47]         0.17    0.00 0.02    0.13    0.16    0.17    0.18
#> prob_infected[48]         0.17    0.00 0.02    0.14    0.16    0.17    0.19
#> prob_infected[49]         0.18    0.00 0.02    0.14    0.16    0.18    0.19
#> prob_infected[50]         0.18    0.00 0.02    0.14    0.16    0.18    0.20
#> prob_infected[51]         0.18    0.00 0.02    0.14    0.17    0.18    0.20
#> prob_infected[52]         0.19    0.00 0.02    0.15    0.17    0.19    0.20
#> prob_infected[53]         0.19    0.00 0.02    0.15    0.17    0.19    0.21
#> prob_infected[54]         0.19    0.00 0.02    0.15    0.18    0.19    0.21
#> prob_infected[55]         0.20    0.00 0.02    0.15    0.18    0.20    0.21
#> prob_infected[56]         0.20    0.00 0.03    0.16    0.18    0.20    0.22
#> prob_infected[57]         0.20    0.00 0.03    0.16    0.18    0.20    0.22
#> prob_infected[58]         0.21    0.00 0.03    0.16    0.19    0.21    0.22
#> prob_infected[59]         0.21    0.00 0.03    0.16    0.19    0.21    0.23
#> prob_infected[60]         0.21    0.00 0.03    0.17    0.19    0.21    0.23
#> prob_infected[61]         0.22    0.00 0.03    0.17    0.20    0.21    0.23
#> prob_infected[62]         0.22    0.00 0.03    0.17    0.20    0.22    0.24
#> prob_infected[63]         0.22    0.00 0.03    0.17    0.20    0.22    0.24
#> prob_infected[64]         0.22    0.00 0.03    0.18    0.20    0.22    0.24
#> prob_infected[65]         0.23    0.00 0.03    0.18    0.21    0.23    0.24
#> prob_infected[66]         0.23    0.00 0.03    0.18    0.21    0.23    0.25
#> prob_infected[67]         0.23    0.00 0.03    0.18    0.21    0.23    0.25
#> prob_infected[68]         0.24    0.00 0.03    0.19    0.22    0.24    0.25
#> prob_infected[69]         0.24    0.00 0.03    0.19    0.22    0.24    0.26
#> prob_infected[70]         0.24    0.00 0.03    0.19    0.22    0.24    0.26
#> prob_infected[71]         0.25    0.00 0.03    0.19    0.22    0.24    0.26
#> prob_infected[72]         0.26    0.00 0.03    0.20    0.24    0.26    0.28
#> scalar_dot_product[1]     0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> scalar_dot_product[2]     0.01    0.00 0.00    0.01    0.01    0.01    0.01
#> scalar_dot_product[3]     0.01    0.00 0.00    0.01    0.01    0.01    0.01
#> scalar_dot_product[4]     0.02    0.00 0.00    0.01    0.01    0.02    0.02
#> scalar_dot_product[5]     0.02    0.00 0.00    0.01    0.02    0.02    0.02
#> scalar_dot_product[6]     0.02    0.00 0.00    0.02    0.02    0.02    0.03
#> scalar_dot_product[7]     0.03    0.00 0.00    0.02    0.02    0.03    0.03
#> scalar_dot_product[8]     0.03    0.00 0.00    0.02    0.03    0.03    0.03
#> scalar_dot_product[9]     0.04    0.00 0.01    0.03    0.03    0.04    0.04
#> scalar_dot_product[10]    0.04    0.00 0.01    0.03    0.04    0.04    0.04
#> scalar_dot_product[11]    0.04    0.00 0.01    0.03    0.04    0.04    0.05
#> scalar_dot_product[12]    0.05    0.00 0.01    0.04    0.04    0.05    0.05
#> scalar_dot_product[13]    0.05    0.00 0.01    0.04    0.05    0.05    0.06
#> scalar_dot_product[14]    0.05    0.00 0.01    0.04    0.05    0.05    0.06
#> scalar_dot_product[15]    0.06    0.00 0.01    0.04    0.05    0.06    0.06
#> scalar_dot_product[16]    0.06    0.00 0.01    0.05    0.06    0.06    0.07
#> scalar_dot_product[17]    0.07    0.00 0.01    0.05    0.06    0.07    0.07
#> scalar_dot_product[18]    0.07    0.00 0.01    0.05    0.06    0.07    0.08
#> scalar_dot_product[19]    0.07    0.00 0.01    0.06    0.07    0.07    0.08
#> scalar_dot_product[20]    0.08    0.00 0.01    0.06    0.07    0.08    0.09
#> scalar_dot_product[21]    0.08    0.00 0.01    0.06    0.07    0.08    0.09
#> scalar_dot_product[22]    0.09    0.00 0.01    0.07    0.08    0.09    0.09
#> scalar_dot_product[23]    0.09    0.00 0.01    0.07    0.08    0.09    0.10
#> scalar_dot_product[24]    0.09    0.00 0.01    0.07    0.08    0.09    0.10
#> scalar_dot_product[25]    0.10    0.00 0.01    0.07    0.09    0.10    0.11
#> scalar_dot_product[26]    0.10    0.00 0.01    0.08    0.09    0.10    0.11
#> scalar_dot_product[27]    0.11    0.00 0.02    0.08    0.09    0.11    0.11
#> scalar_dot_product[28]    0.11    0.00 0.02    0.08    0.10    0.11    0.12
#> scalar_dot_product[29]    0.11    0.00 0.02    0.09    0.10    0.11    0.12
#> scalar_dot_product[30]    0.12    0.00 0.02    0.09    0.11    0.12    0.13
#> scalar_dot_product[31]    0.12    0.00 0.02    0.09    0.11    0.12    0.13
#> scalar_dot_product[32]    0.13    0.00 0.02    0.10    0.11    0.12    0.14
#> scalar_dot_product[33]    0.13    0.00 0.02    0.10    0.12    0.13    0.14
#> scalar_dot_product[34]    0.13    0.00 0.02    0.10    0.12    0.13    0.14
#> scalar_dot_product[35]    0.14    0.00 0.02    0.10    0.12    0.14    0.15
#> scalar_dot_product[36]    0.14    0.00 0.02    0.11    0.13    0.14    0.15
#> scalar_dot_product[37]    0.15    0.00 0.02    0.11    0.13    0.14    0.16
#> scalar_dot_product[38]    0.15    0.00 0.02    0.11    0.13    0.15    0.16
#> scalar_dot_product[39]    0.15    0.00 0.02    0.12    0.14    0.15    0.17
#> scalar_dot_product[40]    0.16    0.00 0.02    0.12    0.14    0.16    0.17
#> scalar_dot_product[41]    0.16    0.00 0.02    0.12    0.14    0.16    0.17
#> scalar_dot_product[42]    0.16    0.00 0.02    0.12    0.15    0.16    0.18
#> scalar_dot_product[43]    0.17    0.00 0.02    0.13    0.15    0.17    0.18
#> scalar_dot_product[44]    0.17    0.00 0.02    0.13    0.15    0.17    0.19
#> scalar_dot_product[45]    0.18    0.00 0.03    0.13    0.16    0.18    0.19
#> scalar_dot_product[46]    0.18    0.00 0.03    0.14    0.16    0.18    0.20
#> scalar_dot_product[47]    0.19    0.00 0.03    0.14    0.17    0.19    0.20
#> scalar_dot_product[48]    0.19    0.00 0.03    0.15    0.17    0.19    0.21
#> scalar_dot_product[49]    0.20    0.00 0.03    0.15    0.18    0.20    0.21
#> scalar_dot_product[50]    0.20    0.00 0.03    0.15    0.18    0.20    0.22
#> scalar_dot_product[51]    0.20    0.00 0.03    0.15    0.18    0.20    0.22
#> scalar_dot_product[52]    0.21    0.00 0.03    0.16    0.19    0.21    0.23
#> scalar_dot_product[53]    0.21    0.00 0.03    0.16    0.19    0.21    0.23
#> scalar_dot_product[54]    0.22    0.00 0.03    0.16    0.19    0.21    0.23
#> scalar_dot_product[55]    0.22    0.00 0.03    0.17    0.20    0.22    0.24
#> scalar_dot_product[56]    0.22    0.00 0.03    0.17    0.20    0.22    0.24
#> scalar_dot_product[57]    0.23    0.00 0.03    0.17    0.20    0.23    0.25
#> scalar_dot_product[58]    0.23    0.00 0.03    0.18    0.21    0.23    0.25
#> scalar_dot_product[59]    0.24    0.00 0.03    0.18    0.21    0.23    0.26
#> scalar_dot_product[60]    0.24    0.00 0.03    0.18    0.21    0.24    0.26
#> scalar_dot_product[61]    0.24    0.00 0.03    0.18    0.22    0.24    0.26
#> scalar_dot_product[62]    0.25    0.00 0.04    0.19    0.22    0.25    0.27
#> scalar_dot_product[63]    0.25    0.00 0.04    0.19    0.23    0.25    0.27
#> scalar_dot_product[64]    0.26    0.00 0.04    0.19    0.23    0.25    0.28
#> scalar_dot_product[65]    0.26    0.00 0.04    0.20    0.23    0.26    0.28
#> scalar_dot_product[66]    0.26    0.00 0.04    0.20    0.24    0.26    0.29
#> scalar_dot_product[67]    0.27    0.00 0.04    0.20    0.24    0.27    0.29
#> scalar_dot_product[68]    0.27    0.00 0.04    0.20    0.24    0.27    0.29
#> scalar_dot_product[69]    0.27    0.00 0.04    0.21    0.25    0.27    0.30
#> scalar_dot_product[70]    0.28    0.00 0.04    0.21    0.25    0.28    0.30
#> scalar_dot_product[71]    0.28    0.00 0.04    0.21    0.25    0.28    0.31
#> scalar_dot_product[72]    0.30    0.00 0.04    0.23    0.27    0.30    0.33
#> foi[1]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[2]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[3]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[4]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[5]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[6]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[7]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[8]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[9]                    0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[10]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[11]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[12]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[13]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[14]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[15]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[16]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[17]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[18]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[19]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[20]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[21]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[22]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[23]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[24]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[25]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[26]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[27]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[28]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[29]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[30]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[31]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[32]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[33]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[34]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[35]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[36]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[37]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[38]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[39]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[40]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[41]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[42]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[43]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[44]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[45]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[46]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[47]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[48]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[49]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[50]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[51]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[52]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[53]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[54]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[55]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[56]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[57]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[58]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[59]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[60]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[61]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[62]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[63]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[64]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[65]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[66]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[67]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[68]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[69]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[70]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[71]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[72]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[73]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[74]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[75]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[76]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> foi[77]                   0.00    0.00 0.00    0.00    0.00    0.00    0.00
#> n_pos_sim[1]              0.13    0.01 0.36    0.00    0.00    0.00    0.00
#> n_pos_sim[2]              0.19    0.01 0.45    0.00    0.00    0.00    0.00
#> n_pos_sim[3]              0.36    0.02 0.59    0.00    0.00    0.00    1.00
#> n_pos_sim[4]              0.45    0.02 0.67    0.00    0.00    0.00    1.00
#> n_pos_sim[5]              0.68    0.02 0.81    0.00    0.00    0.50    1.00
#> n_pos_sim[6]              0.52    0.02 0.68    0.00    0.00    0.00    1.00
#> n_pos_sim[7]              0.91    0.03 0.91    0.00    0.00    1.00    1.00
#> n_pos_sim[8]              0.58    0.02 0.76    0.00    0.00    0.00    1.00
#> n_pos_sim[9]              0.69    0.03 0.82    0.00    0.00    1.00    1.00
#> n_pos_sim[10]             0.95    0.03 0.94    0.00    0.00    1.00    1.00
#> n_pos_sim[11]             0.58    0.02 0.78    0.00    0.00    0.00    1.00
#> n_pos_sim[12]             1.30    0.04 1.16    0.00    0.00    1.00    2.00
#> n_pos_sim[13]             0.72    0.03 0.84    0.00    0.00    1.00    1.00
#> n_pos_sim[14]             1.37    0.04 1.14    0.00    1.00    1.00    2.00
#> n_pos_sim[15]             0.78    0.03 0.87    0.00    0.00    1.00    1.00
#> n_pos_sim[16]             1.00    0.03 0.95    0.00    0.00    1.00    2.00
#> n_pos_sim[17]             0.77    0.03 0.86    0.00    0.00    1.00    1.00
#> n_pos_sim[18]             1.24    0.03 1.04    0.00    0.00    1.00    2.00
#> n_pos_sim[19]             0.64    0.02 0.78    0.00    0.00    0.00    1.00
#> n_pos_sim[20]             0.59    0.02 0.75    0.00    0.00    0.00    1.00
#> n_pos_sim[21]             0.33    0.02 0.55    0.00    0.00    0.00    1.00
#> n_pos_sim[22]             1.58    0.04 1.20    0.00    1.00    1.00    2.00
#> n_pos_sim[23]             1.30    0.03 1.08    0.00    0.00    1.00    2.00
#> n_pos_sim[24]             1.30    0.04 1.09    0.00    0.00    1.00    2.00
#> n_pos_sim[25]             1.89    0.04 1.29    0.00    1.00    2.00    3.00
#> n_pos_sim[26]             1.07    0.03 0.99    0.00    0.00    1.00    2.00
#> n_pos_sim[27]             0.79    0.03 0.83    0.00    0.00    1.00    1.00
#> n_pos_sim[28]             0.94    0.03 0.94    0.00    0.00    1.00    1.00
#> n_pos_sim[29]             0.96    0.03 0.89    0.00    0.00    1.00    1.00
#> n_pos_sim[30]             1.10    0.03 0.99    0.00    0.00    1.00    2.00
#> n_pos_sim[31]             1.07    0.03 0.94    0.00    0.00    1.00    2.00
#> n_pos_sim[32]             1.27    0.04 1.11    0.00    0.00    1.00    2.00
#> n_pos_sim[33]             1.37    0.03 1.10    0.00    1.00    1.00    2.00
#> n_pos_sim[34]             0.97    0.03 0.94    0.00    0.00    1.00    2.00
#> n_pos_sim[35]             1.30    0.03 1.05    0.00    0.00    1.00    2.00
#> n_pos_sim[36]             1.37    0.03 1.08    0.00    1.00    1.00    2.00
#> n_pos_sim[37]             1.24    0.03 1.02    0.00    0.00    1.00    2.00
#> n_pos_sim[38]             0.57    0.03 0.67    0.00    0.00    0.00    1.00
#> n_pos_sim[39]             0.58    0.02 0.70    0.00    0.00    0.00    1.00
#> n_pos_sim[40]             1.30    0.03 1.04    0.00    1.00    1.00    2.00
#> n_pos_sim[41]             0.61    0.02 0.71    0.00    0.00    0.00    1.00
#> n_pos_sim[42]             1.20    0.03 0.99    0.00    0.00    1.00    2.00
#> n_pos_sim[43]             0.46    0.02 0.65    0.00    0.00    0.00    1.00
#> n_pos_sim[44]             0.17    0.01 0.37    0.00    0.00    0.00    0.00
#> n_pos_sim[45]             0.45    0.02 0.60    0.00    0.00    0.00    1.00
#> n_pos_sim[46]             0.51    0.02 0.66    0.00    0.00    0.00    1.00
#> n_pos_sim[47]             0.34    0.02 0.54    0.00    0.00    0.00    1.00
#> n_pos_sim[48]             0.54    0.02 0.65    0.00    0.00    0.00    1.00
#> n_pos_sim[49]             0.56    0.02 0.68    0.00    0.00    0.00    1.00
#> n_pos_sim[50]             0.18    0.01 0.39    0.00    0.00    0.00    0.00
#> n_pos_sim[51]             1.15    0.03 0.98    0.00    0.00    1.00    2.00
#> n_pos_sim[52]             0.73    0.03 0.78    0.00    0.00    1.00    1.00
#> n_pos_sim[53]             0.53    0.02 0.66    0.00    0.00    0.00    1.00
#> n_pos_sim[54]             1.09    0.03 0.95    0.00    0.00    1.00    2.00
#> n_pos_sim[55]             0.21    0.01 0.41    0.00    0.00    0.00    0.00
#> n_pos_sim[56]             0.82    0.03 0.80    0.00    0.00    1.00    1.00
#> n_pos_sim[57]             0.62    0.02 0.71    0.00    0.00    0.00    1.00
#> n_pos_sim[58]             0.19    0.01 0.39    0.00    0.00    0.00    0.00
#> n_pos_sim[59]             2.25    0.05 1.40    0.00    1.00    2.00    3.00
#> n_pos_sim[60]             1.07    0.03 0.93    0.00    0.00    1.00    2.00
#> n_pos_sim[61]             0.20    0.01 0.40    0.00    0.00    0.00    0.00
#> n_pos_sim[62]             0.45    0.02 0.58    0.00    0.00    0.00    1.00
#> n_pos_sim[63]             1.12    0.03 0.95    0.00    0.00    1.00    2.00
#> n_pos_sim[64]             1.84    0.04 1.20    0.00    1.00    2.00    3.00
#> n_pos_sim[65]             0.22    0.01 0.42    0.00    0.00    0.00    0.00
#> n_pos_sim[66]             0.23    0.01 0.42    0.00    0.00    0.00    0.00
#> n_pos_sim[67]             0.46    0.02 0.61    0.00    0.00    0.00    1.00
#> n_pos_sim[68]             1.19    0.03 0.96    0.00    0.00    1.00    2.00
#> n_pos_sim[69]             0.72    0.02 0.71    0.00    0.00    1.00    1.00
#> n_pos_sim[70]             0.24    0.01 0.43    0.00    0.00    0.00    0.00
#> n_pos_sim[71]             0.25    0.01 0.43    0.00    0.00    0.00    0.00
#> n_pos_sim[72]             0.27    0.01 0.45    0.00    0.00    0.00    1.00
#> P_sim[1]                  0.00    0.00 0.01    0.00    0.00    0.00    0.00
#> P_sim[2]                  0.01    0.00 0.02    0.00    0.00    0.00    0.00
#> P_sim[3]                  0.01    0.00 0.02    0.00    0.00    0.00    0.03
#> P_sim[4]                  0.02    0.00 0.02    0.00    0.00    0.00    0.03
#> P_sim[5]                  0.02    0.00 0.02    0.00    0.00    0.01    0.03
#> P_sim[6]                  0.02    0.00 0.03    0.00    0.00    0.00    0.04
#> P_sim[7]                  0.03    0.00 0.03    0.00    0.00    0.03    0.03
#> P_sim[8]                  0.03    0.00 0.04    0.00    0.00    0.00    0.06
#> P_sim[9]                  0.03    0.00 0.04    0.00    0.00    0.05    0.05
#> P_sim[10]                 0.04    0.00 0.04    0.00    0.00    0.04    0.04
#> P_sim[11]                 0.04    0.00 0.06    0.00    0.00    0.00    0.08
#> P_sim[12]                 0.05    0.00 0.04    0.00    0.00    0.04    0.07
#> P_sim[13]                 0.05    0.00 0.06    0.00    0.00    0.07    0.07
#> P_sim[14]                 0.05    0.00 0.05    0.00    0.04    0.04    0.08
#> P_sim[15]                 0.06    0.00 0.07    0.00    0.00    0.08    0.08
#> P_sim[16]                 0.06    0.00 0.06    0.00    0.00    0.06    0.12
#> P_sim[17]                 0.06    0.00 0.07    0.00    0.00    0.08    0.08
#> P_sim[18]                 0.07    0.00 0.06    0.00    0.00    0.06    0.11
#> P_sim[19]                 0.07    0.00 0.09    0.00    0.00    0.00    0.11
#> P_sim[20]                 0.07    0.00 0.09    0.00    0.00    0.00    0.12
#> P_sim[21]                 0.08    0.00 0.14    0.00    0.00    0.00    0.25
#> P_sim[22]                 0.08    0.00 0.06    0.00    0.05    0.05    0.11
#> P_sim[23]                 0.09    0.00 0.07    0.00    0.00    0.07    0.13
#> P_sim[24]                 0.09    0.00 0.08    0.00    0.00    0.07    0.14
#> P_sim[25]                 0.09    0.00 0.06    0.00    0.05    0.10    0.15
#> P_sim[26]                 0.10    0.00 0.09    0.00    0.00    0.09    0.18
#> P_sim[27]                 0.10    0.00 0.10    0.00    0.00    0.12    0.12
#> P_sim[28]                 0.10    0.00 0.10    0.00    0.00    0.11    0.11
#> P_sim[29]                 0.11    0.00 0.10    0.00    0.00    0.11    0.11
#> P_sim[30]                 0.11    0.00 0.10    0.00    0.00    0.10    0.20
#> P_sim[31]                 0.11    0.00 0.09    0.00    0.00    0.10    0.20
#> P_sim[32]                 0.12    0.00 0.10    0.00    0.00    0.09    0.18
#> P_sim[33]                 0.12    0.00 0.10    0.00    0.09    0.09    0.18
#> P_sim[34]                 0.12    0.00 0.12    0.00    0.00    0.12    0.25
#> P_sim[35]                 0.13    0.00 0.11    0.00    0.00    0.10    0.20
#> P_sim[36]                 0.14    0.00 0.11    0.00    0.10    0.10    0.20
#> P_sim[37]                 0.14    0.00 0.11    0.00    0.00    0.11    0.22
#> P_sim[38]                 0.14    0.01 0.17    0.00    0.00    0.00    0.25
#> P_sim[39]                 0.14    0.01 0.18    0.00    0.00    0.00    0.25
#> P_sim[40]                 0.14    0.00 0.12    0.00    0.11    0.11    0.22
#> P_sim[41]                 0.15    0.01 0.18    0.00    0.00    0.00    0.25
#> P_sim[42]                 0.15    0.00 0.12    0.00    0.00    0.12    0.25
#> P_sim[43]                 0.15    0.01 0.22    0.00    0.00    0.00    0.33
#> P_sim[44]                 0.17    0.01 0.37    0.00    0.00    0.00    0.00
#> P_sim[45]                 0.15    0.01 0.20    0.00    0.00    0.00    0.33
#> P_sim[46]                 0.17    0.01 0.22    0.00    0.00    0.00    0.33
#> P_sim[47]                 0.17    0.01 0.27    0.00    0.00    0.00    0.50
#> P_sim[48]                 0.18    0.01 0.22    0.00    0.00    0.00    0.33
#> P_sim[49]                 0.19    0.01 0.23    0.00    0.00    0.00    0.33
#> P_sim[50]                 0.18    0.01 0.39    0.00    0.00    0.00    0.00
#> P_sim[51]                 0.19    0.01 0.16    0.00    0.00    0.17    0.33
#> P_sim[52]                 0.18    0.01 0.19    0.00    0.00    0.25    0.25
#> P_sim[53]                 0.18    0.01 0.22    0.00    0.00    0.00    0.33
#> P_sim[54]                 0.18    0.00 0.16    0.00    0.00    0.17    0.33
#> P_sim[55]                 0.21    0.01 0.41    0.00    0.00    0.00    0.00
#> P_sim[56]                 0.20    0.01 0.20    0.00    0.00    0.25    0.25
#> P_sim[57]                 0.21    0.01 0.24    0.00    0.00    0.00    0.33
#> P_sim[58]                 0.19    0.01 0.39    0.00    0.00    0.00    0.00
#> P_sim[59]                 0.20    0.00 0.13    0.00    0.09    0.18    0.27
#> P_sim[60]                 0.21    0.01 0.19    0.00    0.00    0.20    0.40
#> P_sim[61]                 0.20    0.01 0.40    0.00    0.00    0.00    0.00
#> P_sim[62]                 0.23    0.01 0.29    0.00    0.00    0.00    0.50
#> P_sim[63]                 0.22    0.01 0.19    0.00    0.00    0.20    0.40
#> P_sim[64]                 0.23    0.00 0.15    0.00    0.12    0.25    0.38
#> P_sim[65]                 0.22    0.01 0.42    0.00    0.00    0.00    0.00
#> P_sim[66]                 0.23    0.01 0.42    0.00    0.00    0.00    0.00
#> P_sim[67]                 0.23    0.01 0.30    0.00    0.00    0.00    0.50
#> P_sim[68]                 0.24    0.01 0.19    0.00    0.00    0.20    0.40
#> P_sim[69]                 0.24    0.01 0.24    0.00    0.00    0.33    0.33
#> P_sim[70]                 0.24    0.01 0.43    0.00    0.00    0.00    0.00
#> P_sim[71]                 0.25    0.01 0.43    0.00    0.00    0.00    0.00
#> P_sim[72]                 0.27    0.01 0.45    0.00    0.00    0.00    1.00
#> logLikelihood[1]         -0.13    0.00 0.02   -0.17   -0.14   -0.13   -0.12
#> logLikelihood[2]         -0.20    0.00 0.03   -0.25   -0.21   -0.20   -0.18
#> logLikelihood[3]         -1.30    0.00 0.08   -1.47   -1.36   -1.30   -1.25
#> logLikelihood[4]         -0.46    0.00 0.06   -0.59   -0.49   -0.45   -0.41
#> logLikelihood[5]         -0.71    0.00 0.10   -0.91   -0.77   -0.70   -0.63
#> logLikelihood[6]         -0.54    0.00 0.08   -0.70   -0.59   -0.54   -0.49
#> logLikelihood[7]         -0.93    0.01 0.13   -1.20   -1.01   -0.93   -0.84
#> logLikelihood[8]         -0.56    0.00 0.08   -0.73   -0.61   -0.56   -0.51
#> logLikelihood[9]         -0.74    0.00 0.11   -0.96   -0.80   -0.74   -0.66
#> logLikelihood[10]        -0.94    0.01 0.13   -1.21   -1.02   -0.94   -0.84
#> logLikelihood[11]        -1.13    0.00 0.06   -1.26   -1.17   -1.12   -1.08
#> logLikelihood[12]        -1.32    0.01 0.19   -1.70   -1.43   -1.31   -1.18
#> logLikelihood[13]        -0.77    0.00 0.11   -0.99   -0.83   -0.76   -0.69
#> logLikelihood[14]        -1.37    0.01 0.20   -1.77   -1.49   -1.37   -1.23
#> logLikelihood[15]        -0.77    0.00 0.11   -0.99   -0.83   -0.76   -0.69
#> logLikelihood[16]        -1.00    0.01 0.14   -1.29   -1.09   -1.00   -0.90
#> logLikelihood[17]        -0.80    0.00 0.11   -1.03   -0.87   -0.80   -0.72
#> logLikelihood[18]        -1.27    0.01 0.18   -1.64   -1.38   -1.26   -1.14
#> logLikelihood[19]        -0.67    0.00 0.10   -0.86   -0.73   -0.67   -0.60
#> logLikelihood[20]        -0.63    0.00 0.09   -0.81   -0.68   -0.62   -0.56
#> logLikelihood[21]        -0.33    0.00 0.05   -0.42   -0.36   -0.33   -0.30
#> logLikelihood[22]        -1.64    0.01 0.23   -2.11   -1.78   -1.63   -1.47
#> logLikelihood[23]        -1.35    0.01 0.19   -1.74   -1.47   -1.35   -1.21
#> logLikelihood[24]        -1.32    0.01 0.19   -1.70   -1.43   -1.31   -1.18
#> logLikelihood[25]        -1.96    0.01 0.28   -2.53   -2.13   -1.95   -1.76
#> logLikelihood[26]        -0.97    0.00 0.02   -1.01   -0.97   -0.96   -0.95
#> logLikelihood[27]        -0.97    0.00 0.03   -1.05   -0.99   -0.96   -0.95
#> logLikelihood[28]        -0.99    0.01 0.14   -1.27   -1.07   -0.98   -0.89
#> logLikelihood[29]        -1.02    0.01 0.15   -1.32   -1.11   -1.02   -0.92
#> logLikelihood[30]        -0.96    0.00 0.02   -1.02   -0.97   -0.96   -0.95
#> logLikelihood[31]        -0.97    0.00 0.03   -1.04   -0.98   -0.96   -0.95
#> logLikelihood[32]        -1.38    0.01 0.20   -1.78   -1.50   -1.37   -1.24
#> logLikelihood[33]        -1.02    0.00 0.05   -1.14   -1.04   -1.00   -0.97
#> logLikelihood[34]        -1.07    0.01 0.15   -1.38   -1.16   -1.06   -0.96
#> logLikelihood[35]        -1.00    0.00 0.05   -1.11   -1.02   -0.98   -0.96
#> logLikelihood[36]        -2.31    0.01 0.25   -2.83   -2.49   -2.29   -2.14
#> logLikelihood[37]        -0.98    0.00 0.04   -1.07   -0.99   -0.96   -0.95
#> logLikelihood[38]        -1.05    0.00 0.07   -1.19   -1.09   -1.04   -1.00
#> logLikelihood[39]        -0.61    0.00 0.09   -0.79   -0.66   -0.61   -0.55
#> logLikelihood[40]        -1.39    0.00 0.11   -1.63   -1.46   -1.38   -1.32
#> logLikelihood[41]        -2.36    0.01 0.21   -2.78   -2.51   -2.35   -2.22
#> logLikelihood[42]        -1.32    0.01 0.19   -1.70   -1.43   -1.31   -1.18
#> logLikelihood[43]        -1.11    0.00 0.08   -1.28   -1.17   -1.10   -1.06
#> logLikelihood[44]        -1.85    0.01 0.13   -2.10   -1.94   -1.85   -1.77
#> logLikelihood[45]        -0.53    0.00 0.08   -0.68   -0.57   -0.53   -0.47
#> logLikelihood[46]        -2.70    0.01 0.23   -3.15   -2.87   -2.69   -2.55
#> logLikelihood[47]        -3.54    0.01 0.26   -4.04   -3.72   -3.54   -3.38
#> logLikelihood[48]        -2.60    0.01 0.23   -3.05   -2.76   -2.59   -2.45
#> logLikelihood[49]        -0.59    0.00 0.08   -0.76   -0.64   -0.59   -0.53
#> logLikelihood[50]        -1.72    0.01 0.13   -1.96   -1.81   -1.71   -1.63
#> logLikelihood[51]        -2.72    0.01 0.29   -3.30   -2.92   -2.70   -2.52
#> logLikelihood[52]        -0.83    0.00 0.12   -1.07   -0.90   -0.83   -0.75
#> logLikelihood[53]        -0.99    0.00 0.07   -1.13   -1.04   -0.98   -0.94
#> logLikelihood[54]        -2.60    0.01 0.29   -3.17   -2.80   -2.58   -2.41
#> logLikelihood[55]        -0.22    0.00 0.03   -0.28   -0.24   -0.22   -0.20
#> logLikelihood[56]        -3.69    0.01 0.35   -4.36   -3.93   -3.67   -3.46
#> logLikelihood[57]        -2.33    0.01 0.22   -2.76   -2.49   -2.32   -2.19
#> logLikelihood[58]        -0.23    0.00 0.03   -0.30   -0.25   -0.23   -0.21
#> logLikelihood[59]        -1.53    0.01 0.21   -1.98   -1.65   -1.51   -1.37
#> logLikelihood[60]        -1.20    0.01 0.17   -1.54   -1.30   -1.19   -1.07
#> logLikelihood[61]        -1.54    0.01 0.12   -1.78   -1.63   -1.54   -1.46
#> logLikelihood[62]        -1.08    0.00 0.09   -1.26   -1.14   -1.08   -1.02
#> logLikelihood[63]        -2.74    0.01 0.30   -3.34   -2.95   -2.73   -2.54
#> logLikelihood[64]        -1.20    0.00 0.04   -1.31   -1.22   -1.18   -1.17
#> logLikelihood[65]        -1.49    0.00 0.12   -1.73   -1.57   -1.48   -1.41
#> logLikelihood[66]        -1.47    0.00 0.12   -1.71   -1.56   -1.47   -1.39
#> logLikelihood[67]        -0.53    0.00 0.08   -0.69   -0.58   -0.53   -0.48
#> logLikelihood[68]        -1.41    0.01 0.13   -1.68   -1.50   -1.39   -1.32
#> logLikelihood[69]        -2.05    0.01 0.21   -2.46   -2.19   -2.04   -1.91
#> logLikelihood[70]        -1.42    0.00 0.12   -1.66   -1.51   -1.42   -1.34
#> logLikelihood[71]        -1.41    0.00 0.12   -1.65   -1.50   -1.41   -1.33
#> logLikelihood[72]        -1.35    0.00 0.12   -1.59   -1.44   -1.35   -1.27
#> lp__                   -156.10    0.04 0.78 -158.06 -156.30 -155.81 -155.60
#>                          97.5% n_eff Rhat
#> lambda0                   0.01   608 1.00
#> prob_infected[1]          0.01   608 1.00
#> prob_infected[2]          0.01   608 1.00
#> prob_infected[3]          0.02   608 1.00
#> prob_infected[4]          0.02   608 1.00
#> prob_infected[5]          0.02   609 1.00
#> prob_infected[6]          0.03   609 1.00
#> prob_infected[7]          0.03   609 1.00
#> prob_infected[8]          0.04   609 1.00
#> prob_infected[9]          0.04   609 1.00
#> prob_infected[10]         0.05   609 1.00
#> prob_infected[11]         0.05   609 1.00
#> prob_infected[12]         0.06   609 1.00
#> prob_infected[13]         0.06   609 1.00
#> prob_infected[14]         0.07   609 1.00
#> prob_infected[15]         0.07   609 1.00
#> prob_infected[16]         0.08   609 1.00
#> prob_infected[17]         0.08   609 1.00
#> prob_infected[18]         0.09   609 1.00
#> prob_infected[19]         0.09   609 1.00
#> prob_infected[20]         0.10   609 1.00
#> prob_infected[21]         0.10   609 1.00
#> prob_infected[22]         0.11   609 1.00
#> prob_infected[23]         0.11   610 1.00
#> prob_infected[24]         0.11   610 1.00
#> prob_infected[25]         0.12   610 1.00
#> prob_infected[26]         0.12   610 1.00
#> prob_infected[27]         0.13   610 1.00
#> prob_infected[28]         0.13   610 1.00
#> prob_infected[29]         0.14   610 1.00
#> prob_infected[30]         0.14   610 1.00
#> prob_infected[31]         0.15   610 1.00
#> prob_infected[32]         0.15   610 1.00
#> prob_infected[33]         0.15   610 1.00
#> prob_infected[34]         0.16   610 1.00
#> prob_infected[35]         0.16   610 1.00
#> prob_infected[36]         0.17   610 1.00
#> prob_infected[37]         0.17   610 1.00
#> prob_infected[38]         0.17   610 1.00
#> prob_infected[39]         0.18   610 1.00
#> prob_infected[40]         0.18   610 1.00
#> prob_infected[41]         0.19   610 1.00
#> prob_infected[42]         0.19   611 1.00
#> prob_infected[43]         0.20   611 1.00
#> prob_infected[44]         0.20   611 1.00
#> prob_infected[45]         0.20   611 1.00
#> prob_infected[46]         0.21   611 1.00
#> prob_infected[47]         0.22   611 1.00
#> prob_infected[48]         0.22   611 1.00
#> prob_infected[49]         0.22   611 1.00
#> prob_infected[50]         0.23   611 1.00
#> prob_infected[51]         0.23   611 1.00
#> prob_infected[52]         0.24   611 1.00
#> prob_infected[53]         0.24   611 1.00
#> prob_infected[54]         0.24   611 1.00
#> prob_infected[55]         0.25   611 1.00
#> prob_infected[56]         0.25   611 1.00
#> prob_infected[57]         0.25   611 1.00
#> prob_infected[58]         0.26   611 1.00
#> prob_infected[59]         0.26   611 1.00
#> prob_infected[60]         0.27   611 1.00
#> prob_infected[61]         0.27   612 1.00
#> prob_infected[62]         0.27   612 1.00
#> prob_infected[63]         0.28   612 1.00
#> prob_infected[64]         0.28   612 1.00
#> prob_infected[65]         0.28   612 1.00
#> prob_infected[66]         0.29   612 1.00
#> prob_infected[67]         0.29   612 1.00
#> prob_infected[68]         0.29   612 1.00
#> prob_infected[69]         0.30   612 1.00
#> prob_infected[70]         0.30   612 1.00
#> prob_infected[71]         0.31   612 1.00
#> prob_infected[72]         0.32   612 1.00
#> scalar_dot_product[1]     0.01   608 1.00
#> scalar_dot_product[2]     0.01   608 1.00
#> scalar_dot_product[3]     0.02   608 1.00
#> scalar_dot_product[4]     0.02   608 1.00
#> scalar_dot_product[5]     0.03   608 1.00
#> scalar_dot_product[6]     0.03   608 1.00
#> scalar_dot_product[7]     0.04   608 1.00
#> scalar_dot_product[8]     0.04   608 1.00
#> scalar_dot_product[9]     0.05   608 1.00
#> scalar_dot_product[10]    0.05   608 1.00
#> scalar_dot_product[11]    0.06   608 1.00
#> scalar_dot_product[12]    0.06   608 1.00
#> scalar_dot_product[13]    0.07   608 1.00
#> scalar_dot_product[14]    0.07   608 1.00
#> scalar_dot_product[15]    0.08   608 1.00
#> scalar_dot_product[16]    0.08   608 1.00
#> scalar_dot_product[17]    0.09   608 1.00
#> scalar_dot_product[18]    0.09   608 1.00
#> scalar_dot_product[19]    0.10   608 1.00
#> scalar_dot_product[20]    0.10   608 1.00
#> scalar_dot_product[21]    0.11   608 1.00
#> scalar_dot_product[22]    0.11   608 1.00
#> scalar_dot_product[23]    0.12   608 1.00
#> scalar_dot_product[24]    0.12   608 1.00
#> scalar_dot_product[25]    0.13   608 1.00
#> scalar_dot_product[26]    0.13   608 1.00
#> scalar_dot_product[27]    0.14   608 1.00
#> scalar_dot_product[28]    0.14   608 1.00
#> scalar_dot_product[29]    0.15   608 1.00
#> scalar_dot_product[30]    0.15   608 1.00
#> scalar_dot_product[31]    0.16   608 1.00
#> scalar_dot_product[32]    0.16   608 1.00
#> scalar_dot_product[33]    0.17   608 1.00
#> scalar_dot_product[34]    0.17   608 1.00
#> scalar_dot_product[35]    0.18   608 1.00
#> scalar_dot_product[36]    0.18   608 1.00
#> scalar_dot_product[37]    0.19   608 1.00
#> scalar_dot_product[38]    0.19   608 1.00
#> scalar_dot_product[39]    0.20   608 1.00
#> scalar_dot_product[40]    0.20   608 1.00
#> scalar_dot_product[41]    0.21   608 1.00
#> scalar_dot_product[42]    0.21   608 1.00
#> scalar_dot_product[43]    0.22   608 1.00
#> scalar_dot_product[44]    0.22   608 1.00
#> scalar_dot_product[45]    0.23   608 1.00
#> scalar_dot_product[46]    0.23   608 1.00
#> scalar_dot_product[47]    0.24   608 1.00
#> scalar_dot_product[48]    0.25   608 1.00
#> scalar_dot_product[49]    0.25   608 1.00
#> scalar_dot_product[50]    0.26   608 1.00
#> scalar_dot_product[51]    0.26   608 1.00
#> scalar_dot_product[52]    0.27   608 1.00
#> scalar_dot_product[53]    0.27   608 1.00
#> scalar_dot_product[54]    0.28   608 1.00
#> scalar_dot_product[55]    0.28   608 1.00
#> scalar_dot_product[56]    0.29   608 1.00
#> scalar_dot_product[57]    0.29   608 1.00
#> scalar_dot_product[58]    0.30   608 1.00
#> scalar_dot_product[59]    0.30   608 1.00
#> scalar_dot_product[60]    0.31   608 1.00
#> scalar_dot_product[61]    0.31   608 1.00
#> scalar_dot_product[62]    0.32   608 1.00
#> scalar_dot_product[63]    0.32   608 1.00
#> scalar_dot_product[64]    0.33   608 1.00
#> scalar_dot_product[65]    0.33   608 1.00
#> scalar_dot_product[66]    0.34   608 1.00
#> scalar_dot_product[67]    0.34   608 1.00
#> scalar_dot_product[68]    0.35   608 1.00
#> scalar_dot_product[69]    0.35   608 1.00
#> scalar_dot_product[70]    0.36   608 1.00
#> scalar_dot_product[71]    0.36   608 1.00
#> scalar_dot_product[72]    0.39   608 1.00
#> foi[1]                    0.01   608 1.00
#> foi[2]                    0.01   608 1.00
#> foi[3]                    0.01   608 1.00
#> foi[4]                    0.01   608 1.00
#> foi[5]                    0.01   608 1.00
#> foi[6]                    0.01   608 1.00
#> foi[7]                    0.01   608 1.00
#> foi[8]                    0.01   608 1.00
#> foi[9]                    0.01   608 1.00
#> foi[10]                   0.01   608 1.00
#> foi[11]                   0.01   608 1.00
#> foi[12]                   0.01   608 1.00
#> foi[13]                   0.01   608 1.00
#> foi[14]                   0.01   608 1.00
#> foi[15]                   0.01   608 1.00
#> foi[16]                   0.01   608 1.00
#> foi[17]                   0.01   608 1.00
#> foi[18]                   0.01   608 1.00
#> foi[19]                   0.01   608 1.00
#> foi[20]                   0.01   608 1.00
#> foi[21]                   0.01   608 1.00
#> foi[22]                   0.01   608 1.00
#> foi[23]                   0.01   608 1.00
#> foi[24]                   0.01   608 1.00
#> foi[25]                   0.01   608 1.00
#> foi[26]                   0.01   608 1.00
#> foi[27]                   0.01   608 1.00
#> foi[28]                   0.01   608 1.00
#> foi[29]                   0.01   608 1.00
#> foi[30]                   0.01   608 1.00
#> foi[31]                   0.01   608 1.00
#> foi[32]                   0.01   608 1.00
#> foi[33]                   0.01   608 1.00
#> foi[34]                   0.01   608 1.00
#> foi[35]                   0.01   608 1.00
#> foi[36]                   0.01   608 1.00
#> foi[37]                   0.01   608 1.00
#> foi[38]                   0.01   608 1.00
#> foi[39]                   0.01   608 1.00
#> foi[40]                   0.01   608 1.00
#> foi[41]                   0.01   608 1.00
#> foi[42]                   0.01   608 1.00
#> foi[43]                   0.01   608 1.00
#> foi[44]                   0.01   608 1.00
#> foi[45]                   0.01   608 1.00
#> foi[46]                   0.01   608 1.00
#> foi[47]                   0.01   608 1.00
#> foi[48]                   0.01   608 1.00
#> foi[49]                   0.01   608 1.00
#> foi[50]                   0.01   608 1.00
#> foi[51]                   0.01   608 1.00
#> foi[52]                   0.01   608 1.00
#> foi[53]                   0.01   608 1.00
#> foi[54]                   0.01   608 1.00
#> foi[55]                   0.01   608 1.00
#> foi[56]                   0.01   608 1.00
#> foi[57]                   0.01   608 1.00
#> foi[58]                   0.01   608 1.00
#> foi[59]                   0.01   608 1.00
#> foi[60]                   0.01   608 1.00
#> foi[61]                   0.01   608 1.00
#> foi[62]                   0.01   608 1.00
#> foi[63]                   0.01   608 1.00
#> foi[64]                   0.01   608 1.00
#> foi[65]                   0.01   608 1.00
#> foi[66]                   0.01   608 1.00
#> foi[67]                   0.01   608 1.00
#> foi[68]                   0.01   608 1.00
#> foi[69]                   0.01   608 1.00
#> foi[70]                   0.01   608 1.00
#> foi[71]                   0.01   608 1.00
#> foi[72]                   0.01   608 1.00
#> foi[73]                   0.01   608 1.00
#> foi[74]                   0.01   608 1.00
#> foi[75]                   0.01   608 1.00
#> foi[76]                   0.01   608 1.00
#> foi[77]                   0.01   608 1.00
#> n_pos_sim[1]              1.00  1027 1.00
#> n_pos_sim[2]              1.00   929 1.00
#> n_pos_sim[3]              2.00  1015 1.00
#> n_pos_sim[4]              2.00  1034 1.00
#> n_pos_sim[5]              3.00  1146 1.00
#> n_pos_sim[6]              2.00  1166 1.00
#> n_pos_sim[7]              3.00   939 1.00
#> n_pos_sim[8]              2.00   923 1.00
#> n_pos_sim[9]              3.00   993 1.00
#> n_pos_sim[10]             3.00   842 1.00
#> n_pos_sim[11]             3.00  1052 1.00
#> n_pos_sim[12]             4.00   958 1.00
#> n_pos_sim[13]             3.00  1019 1.00
#> n_pos_sim[14]             4.00   865 1.00
#> n_pos_sim[15]             3.00  1088 1.00
#> n_pos_sim[16]             3.00   938 1.00
#> n_pos_sim[17]             3.00   976 1.00
#> n_pos_sim[18]             4.00   939 1.00
#> n_pos_sim[19]             2.00  1094 1.00
#> n_pos_sim[20]             2.00   955 1.00
#> n_pos_sim[21]             2.00   878 1.00
#> n_pos_sim[22]             4.00   841 1.00
#> n_pos_sim[23]             4.00   968 1.00
#> n_pos_sim[24]             4.00   904 1.00
#> n_pos_sim[25]             5.00   949 1.00
#> n_pos_sim[26]             3.00  1047 1.00
#> n_pos_sim[27]             3.00  1032 1.00
#> n_pos_sim[28]             3.00   952 1.00
#> n_pos_sim[29]             3.00   909 1.00
#> n_pos_sim[30]             3.00   977 1.00
#> n_pos_sim[31]             3.00  1003 1.00
#> n_pos_sim[32]             4.00   870 1.00
#> n_pos_sim[33]             4.00  1017 1.00
#> n_pos_sim[34]             3.00  1012 1.00
#> n_pos_sim[35]             4.00   944 1.00
#> n_pos_sim[36]             4.00   966 1.00
#> n_pos_sim[37]             3.00  1007 1.00
#> n_pos_sim[38]             2.00   715 1.00
#> n_pos_sim[39]             2.00   845 1.01
#> n_pos_sim[40]             4.00   948 1.00
#> n_pos_sim[41]             2.00   899 1.00
#> n_pos_sim[42]             3.00  1072 1.00
#> n_pos_sim[43]             2.00  1007 1.00
#> n_pos_sim[44]             1.00  1067 1.00
#> n_pos_sim[45]             2.00  1058 1.00
#> n_pos_sim[46]             2.00  1105 1.00
#> n_pos_sim[47]             2.00   928 1.00
#> n_pos_sim[48]             2.00  1132 1.00
#> n_pos_sim[49]             2.00  1062 1.00
#> n_pos_sim[50]             1.00  1136 1.00
#> n_pos_sim[51]             3.00   907 1.00
#> n_pos_sim[52]             2.00   929 1.00
#> n_pos_sim[53]             2.00   966 1.00
#> n_pos_sim[54]             3.00  1005 1.00
#> n_pos_sim[55]             1.00   838 1.00
#> n_pos_sim[56]             2.00   950 1.00
#> n_pos_sim[57]             2.00   987 1.00
#> n_pos_sim[58]             1.00   812 1.00
#> n_pos_sim[59]             5.00   960 1.00
#> n_pos_sim[60]             3.00  1134 1.00
#> n_pos_sim[61]             1.00   797 1.00
#> n_pos_sim[62]             2.00   991 1.01
#> n_pos_sim[63]             3.00   951 1.00
#> n_pos_sim[64]             4.00   965 1.00
#> n_pos_sim[65]             1.00   989 1.00
#> n_pos_sim[66]             1.00   932 1.00
#> n_pos_sim[67]             2.00  1058 1.00
#> n_pos_sim[68]             3.00  1131 1.00
#> n_pos_sim[69]             2.00  1020 1.00
#> n_pos_sim[70]             1.00   932 1.00
#> n_pos_sim[71]             1.00  1005 1.00
#> n_pos_sim[72]             1.00   960 1.00
#> P_sim[1]                  0.03  1027 1.00
#> P_sim[2]                  0.04   929 1.00
#> P_sim[3]                  0.06  1015 1.00
#> P_sim[4]                  0.07  1034 1.00
#> P_sim[5]                  0.08  1146 1.00
#> P_sim[6]                  0.09  1166 1.00
#> P_sim[7]                  0.09   939 1.00
#> P_sim[8]                  0.11   923 1.00
#> P_sim[9]                  0.14   993 1.00
#> P_sim[10]                 0.12   842 1.00
#> P_sim[11]                 0.23  1052 1.00
#> P_sim[12]                 0.14   958 1.00
#> P_sim[13]                 0.20  1019 1.00
#> P_sim[14]                 0.16   865 1.00
#> P_sim[15]                 0.23  1088 1.00
#> P_sim[16]                 0.19   938 1.00
#> P_sim[17]                 0.25   976 1.00
#> P_sim[18]                 0.22   939 1.00
#> P_sim[19]                 0.22  1094 1.00
#> P_sim[20]                 0.25   955 1.00
#> P_sim[21]                 0.50   878 1.00
#> P_sim[22]                 0.21   841 1.00
#> P_sim[23]                 0.27   968 1.00
#> P_sim[24]                 0.29   904 1.00
#> P_sim[25]                 0.25   949 1.00
#> P_sim[26]                 0.27  1047 1.00
#> P_sim[27]                 0.38  1032 1.00
#> P_sim[28]                 0.33   952 1.00
#> P_sim[29]                 0.33   909 1.00
#> P_sim[30]                 0.30   977 1.00
#> P_sim[31]                 0.30  1003 1.00
#> P_sim[32]                 0.36   870 1.00
#> P_sim[33]                 0.36  1017 1.00
#> P_sim[34]                 0.38  1012 1.00
#> P_sim[35]                 0.40   944 1.00
#> P_sim[36]                 0.40   966 1.00
#> P_sim[37]                 0.33  1007 1.00
#> P_sim[38]                 0.50   715 1.00
#> P_sim[39]                 0.50   845 1.01
#> P_sim[40]                 0.44   948 1.00
#> P_sim[41]                 0.50   899 1.00
#> P_sim[42]                 0.38  1072 1.00
#> P_sim[43]                 0.67  1007 1.00
#> P_sim[44]                 1.00  1067 1.00
#> P_sim[45]                 0.67  1058 1.00
#> P_sim[46]                 0.67  1105 1.00
#> P_sim[47]                 1.00   928 1.00
#> P_sim[48]                 0.67  1132 1.00
#> P_sim[49]                 0.67  1062 1.00
#> P_sim[50]                 1.00  1136 1.00
#> P_sim[51]                 0.50   907 1.00
#> P_sim[52]                 0.50   929 1.00
#> P_sim[53]                 0.67   966 1.00
#> P_sim[54]                 0.50  1005 1.00
#> P_sim[55]                 1.00   838 1.00
#> P_sim[56]                 0.50   950 1.00
#> P_sim[57]                 0.67   987 1.00
#> P_sim[58]                 1.00   812 1.00
#> P_sim[59]                 0.45   960 1.00
#> P_sim[60]                 0.60  1134 1.00
#> P_sim[61]                 1.00   797 1.00
#> P_sim[62]                 1.00   991 1.01
#> P_sim[63]                 0.60   951 1.00
#> P_sim[64]                 0.50   965 1.00
#> P_sim[65]                 1.00   989 1.00
#> P_sim[66]                 1.00   932 1.00
#> P_sim[67]                 1.00  1058 1.00
#> P_sim[68]                 0.60  1131 1.00
#> P_sim[69]                 0.67  1020 1.00
#> P_sim[70]                 1.00   932 1.00
#> P_sim[71]                 1.00  1005 1.00
#> P_sim[72]                 1.00   960 1.00
#> logLikelihood[1]         -0.10   608 1.00
#> logLikelihood[2]         -0.15   608 1.00
#> logLikelihood[3]         -1.16   620 1.00
#> logLikelihood[4]         -0.34   608 1.00
#> logLikelihood[5]         -0.53   608 1.00
#> logLikelihood[6]         -0.41   608 1.00
#> logLikelihood[7]         -0.71   608 1.00
#> logLikelihood[8]         -0.43   608 1.00
#> logLikelihood[9]         -0.56   608 1.00
#> logLikelihood[10]        -0.71   608 1.00
#> logLikelihood[11]        -1.02   622 1.00
#> logLikelihood[12]        -1.00   608 1.00
#> logLikelihood[13]        -0.58   608 1.00
#> logLikelihood[14]        -1.04   608 1.00
#> logLikelihood[15]        -0.58   608 1.00
#> logLikelihood[16]        -0.76   608 1.00
#> logLikelihood[17]        -0.61   608 1.00
#> logLikelihood[18]        -0.96   608 1.00
#> logLikelihood[19]        -0.51   608 1.00
#> logLikelihood[20]        -0.48   608 1.00
#> logLikelihood[21]        -0.25   608 1.00
#> logLikelihood[22]        -1.24   608 1.00
#> logLikelihood[23]        -1.02   608 1.00
#> logLikelihood[24]        -1.00   608 1.00
#> logLikelihood[25]        -1.48   608 1.00
#> logLikelihood[26]        -0.95   496 1.00
#> logLikelihood[27]        -0.93   617 1.00
#> logLikelihood[28]        -0.75   608 1.00
#> logLikelihood[29]        -0.77   608 1.00
#> logLikelihood[30]        -0.95   521 1.00
#> logLikelihood[31]        -0.95   537 1.00
#> logLikelihood[32]        -1.05   608 1.00
#> logLikelihood[33]        -0.95   576 1.00
#> logLikelihood[34]        -0.81   608 1.00
#> logLikelihood[35]        -0.95   570 1.00
#> logLikelihood[36]        -1.87   620 1.00
#> logLikelihood[37]        -0.94   559 1.00
#> logLikelihood[38]        -0.93   622 1.00
#> logLikelihood[39]        -0.46   608 1.00
#> logLikelihood[40]        -1.23   623 1.00
#> logLikelihood[41]        -1.97   618 1.00
#> logLikelihood[42]        -1.00   608 1.00
#> logLikelihood[43]        -0.97   621 1.00
#> logLikelihood[44]        -1.61   617 1.00
#> logLikelihood[45]        -0.40   608 1.00
#> logLikelihood[46]        -2.28   618 1.00
#> logLikelihood[47]        -3.07   617 1.00
#> logLikelihood[48]        -2.18   618 1.00
#> logLikelihood[49]        -0.45   608 1.00
#> logLikelihood[50]        -1.48   617 1.00
#> logLikelihood[51]        -2.19   619 1.00
#> logLikelihood[52]        -0.63   608 1.00
#> logLikelihood[53]        -0.88   622 1.00
#> logLikelihood[54]        -2.09   619 1.00
#> logLikelihood[55]        -0.17   608 1.00
#> logLikelihood[56]        -3.06   618 1.00
#> logLikelihood[57]        -1.93   618 1.00
#> logLikelihood[58]        -0.18   608 1.00
#> logLikelihood[59]        -1.20   601 1.00
#> logLikelihood[60]        -0.91   608 1.00
#> logLikelihood[61]        -1.31   617 1.00
#> logLikelihood[62]        -0.92   620 1.00
#> logLikelihood[63]        -2.20   619 1.00
#> logLikelihood[64]        -1.17   582 1.00
#> logLikelihood[65]        -1.26   617 1.00
#> logLikelihood[66]        -1.25   617 1.00
#> logLikelihood[67]        -0.40   608 1.00
#> logLikelihood[68]        -1.19   622 1.00
#> logLikelihood[69]        -1.68   619 1.00
#> logLikelihood[70]        -1.20   617 1.00
#> logLikelihood[71]        -1.19   617 1.00
#> logLikelihood[72]        -1.13   618 1.00
#> lp__                   -155.55   473 1.00
#> 
#> Samples were drawn using NUTS(diag_e) at Fri Mar 15 21:25:19 2024.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).